add.cim.covar {qtl}	R Documentation

Indicate marker covariates from composite interval mapping

Description

Add dots at the locations of the selected marker covariates, for a plot of composite interval mapping results.

Usage

add.cim.covar(cimresult, chr, gap=25, ...)

Arguments

cimresult
Composite interval mapping results, as output from cim.

chr
Optional vector specifying which chromosomes to plot. (The chromosomes must be specified by name.) This should be identical to that used in the call to plot.scanone.

gap
Gap separating chromosomes (in cM). This should be identical to that used in the call to plot.scanone.

...
Additional plot arguments, passed to the function points.

Details

One must first have used the function plot.scanone to plot the composite interval mapping results.

The arguments chr and gap must be identical to the values used in the call to plot.scanone.

Dots indicating the locations of the selected marker covariates are displayed on the x-axis. (By default, solid red circles are plotted; this may be modified by specifying the graphics parameters pch and col.)

Value

A data frame indicating the marker covariates that were plotted.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

cim, plot.scanone

Examples

## Not run: data(hyper)
hyper <- calc.genoprob(hyper, step=2.5)

out <- scanone(hyper)
out.cim <- cim(hyper, n.marcovar=3)
plot(out, out.cim, chr=c(1,4,6,15), col=c("blue", "red"))

add.cim.covar(out.cim, chr=c(1,4,6,15))
## End(Not run)

[Package qtl version 1.74 Index]