add.threshold {qtl}	R Documentation

Add significance threshold to plot

Description

Add a significance threshold to a plot created by plot.scanone), using the permutation results.

Usage

add.threshold(out, chr, perms, alpha=0.05, lodcolumn=1, gap=25, ...)

Arguments

out
An object of class "scanone", as output by scanone. This must be identical to what was used in the call to plot.scanone.

chr
Optional vector specifying which chromosomes to plot. If a selected subset of chromosomes were plotted, they must be specified here.

perms
Permutation results from scanone, used to calculate the significance threshold.

alpha
Significance level of the threshold.

lodcolumn
An integer indicating which of column in the permutation results should be used.

gap
Gap separating chromosomes (in cM). This must be identical to what was used in the call to plot.scanone.

...
Passed to the function abline when it is called.

Details

This function allows you to add a horizontal line at the significance threshold to genome scan results plotted by plot.scanone.

The arguments out, chr, and gap must match what was used in the call to plot.scanone.

The argument perms must be specified. If X-chromosome-specific permutations were performed (via the argument perm.Xsp in the call to scanone), separate thresholds will be plotted for the autosomes and the X chromosome. These are calculated via the summary.scanoneperm function.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scanone, plot.scanone, summary.scanoneperm, xaxisloc.scanone

Examples

data(hyper)
hyper <- calc.genoprob(hyper)
out <- scanone(hyper, method="hk")
operm <- scanone(hyper, method="hk", n.perm=100, perm.Xsp=TRUE)

plot(out, chr=c(1,4,6,15,"X"))
add.threshold(out, chr=c(1,4,6,15,"X"), perms=operm, alpha=0.05)
add.threshold(out, chr=c(1,4,6,15,"X"), perms=operm, alpha=0.1,
              col="green", lty=2)

[Package qtl version 1.74 Index]