addloctocross {qtl}	R Documentation

Add phenotype location into a cross object

Description

Add phenotype location(s) into a cross object (with eQTL/pQTL studies)

Usage

    addloctocross(cross, locations=NULL, locfile="locations.txt", verbose=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

locations
R variable holding location information

locfile
load from a file, see the details section for the layout of the file.

verbose
If TRUE, give verbose output

Details

inputfile layout: Num Name Chr cM 1 X3.Hydroxypropyl 4 50.0 Num is the number of the phenotype in the cross object Name is the name of the phenotype (will be checked against the name already in the cross object at position num Chr Chromosome cM position from start of chromosome in cM

Value

The input cross object, with the locations added as an additional component locations

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

mqmplot.cistrans - Cis/trans plot

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

## Not run: 
    data(multitrait)
    data(locations)
    multiloc <- addloctocross(multitrait,locations)
    results <- scanall(multiloc)
    mqmplot.cistrans(results, multiloc, 5, FALSE, TRUE)
  
## End(Not run)

[Package qtl version 1.74 Index]