addtoqtl {qtl}	R Documentation

Add to a qtl object

Description

Add a QTL or multiple QTL to a qtl object.

Usage

addtoqtl(cross, qtl, chr, pos, qtl.name, drop.lod.profile=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

qtl
The qtl object to which additional QTL are to be added.

chr
Vector indicating the chromosome for each new QTL. (These should be character strings referring to the chromosomes by name.)

pos
Vector (of same length as chr) indicating the positions on the chromosome for each new QTL. If there is no marker or pseudomarker at a position, the nearest position is used.

qtl.name
Optional user-specified name for each new QTL, used in the drop-one-term ANOVA table in fitqtl. If unspecified, the names will be of the form "Chr1@10" for a QTL on Chromsome 1 at 10 cM.

drop.lod.profile
If TRUE, remove any LOD profiles from the object.

Value

An object of class qtl, just like the input qtl object, but with additional QTL added. See makeqtl for details.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

makeqtl, fitqtl, dropfromqtl, replaceqtl, reorderqtl

Examples

data(fake.f2)

# take out several QTLs and make QTL object
qc <- c(1, 6, 13)
qp <- c(25.8, 33.6, 18.63)

fake.f2 <- calc.genoprob(fake.f2, step=2, err=0.001)
qtl <- makeqtl(fake.f2, qc, qp, what="prob")
qtl <- addtoqtl(fake.f2, qtl, 14, 35)

[Package qtl version 1.74 Index]