arithscan {qtl}	R Documentation

Arithmetic operators for scanone and scantwo results

Description

Add or subtract LOD scores in results from scanone or scantwo.

Usage

scan1+scan2
scan1-scan2

Arguments

scan1, scan2
Genome scan results on the same set of chromosomes and markers, as output by scanone or scantwo.

Details

This is used to calculate the sum or difference of LOD scores of two genome scan results. It is particularly useful for calculating the LOD scores for QTL-by-covariate interactions (see the example, below). Note that the degrees of freedom are also added or subtracted.

Value

The same type of data structure as the input objects, with LOD scores added or subtracted.

Author(s)

Karl W Broman, broman@wisc.edu

Examples

data(fake.bc)

fake.bc <- calc.genoprob(fake.bc, step=2.5)

# covariates
ac <- pull.pheno(fake.bc, c("sex","age"))
ic <- pull.pheno(fake.bc, "sex")

# scan with additive but not the interactive covariate
out.acovar <- scanone(fake.bc, addcovar=ac)

# scan with interactive covariate
out.icovar <- scanone(fake.bc, addcovar=ac, intcovar=ic)

# plot the difference of with and without the interactive covariate
#     This is a LOD score for a test of QTL x covariate interaction
plot(out.icovar-out.acovar)

[Package qtl version 1.74 Index]