arithscanperm {qtl}	R Documentation

Arithmetic Operators for permutation results

Description

Add or subtract LOD scores in permutation results from scanone or scantwo.

Usage

perm1+perm2
perm1-perm2

Arguments

perm1, perm2
Permutation results from scanone or scantwo, on the same set of chromosomes and markers.

Details

This is used to calculate the sum or difference of LOD scores of two sets of permutation results from scanone or scantwo. One must be careful to ensure that the permutations are perfectly linked, which will require the use of set.seed.

Value

The same data structure as the input objects, with LOD scores added or subtracted.

Author(s)

Karl W Broman, broman@wisc.edu

Examples

data(fake.bc)

fake.bc <- calc.genoprob(fake.bc, step=2.5)

# covariates
ac <- pull.pheno(fake.bc, c("sex","age"))
ic <- pull.pheno(fake.bc, "sex")

# set seed
theseed <- round(runif(1, 1, 10^8))
set.seed(theseed)

# permutations with additive but not the interactive covariate
## Not run: operm.acovar <- scanone(fake.bc, addcovar=ac, n.perm=1000)

# re-set the seed
set.seed(theseed)

# permutations with interactive covariate
## Not run: operm.icovar <- scanone(fake.bc, addcovar=ac, intcovar=ic,
                      n.perm=1000)

## End(Not run)

# permutation results for the QTL x covariate interaction
operm.gxc <- operm.icovar - operm.acovar

# LOD thresholds
summary(operm.gxc)

[Package qtl version 1.74 Index]