bayesint {qtl}	R Documentation

Bayesian credible interval

Description

Calculate an approximate Bayesian credible interval for a particular chromosome, using output from scanone.

Usage

bayesint(results, chr, qtl.index, prob=0.95, lodcolumn=1, expandtomarkers=FALSE)

Arguments

results
Output from scanone, or a qtl object as output from refineqtl.

chr
A chromosome ID (if input results are from scanone (should have length 1).

qtl.index
Numeric index for a QTL (if input results are from refineqtl (should have length 1).

prob
Probability coverage of the interval.

lodcolumn
An integer indicating which of the LOD score columns should be considered (if input results are from scanone).

expandtomarkers
If TRUE, the interval is expanded to the nearest flanking markers.

Details

We take 10^LOD, rescale it to have area 1, and then calculate the connected interval with density above some threshold and having coverage matching the target probability.

Value

An object of class scanone indicating the estimated QTL position and the approximate endpoints for the Bayesian credible interval.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scanone, lodint

Examples

data(hyper)

hyper <- calc.genoprob(hyper, step=0.5)
out <- scanone(hyper, method="hk")
bayesint(out, chr=1)
bayesint(out, chr=4)
bayesint(out, chr=4, prob=0.99)
bayesint(out, chr=4, expandtomarkers=TRUE)

[Package qtl version 1.74 Index]