c.cross {qtl}	R Documentation

Combine data for QTL experiments

Description

Concatenate the data for multiple QTL experiments.

Usage

## S3 method for class 'cross'
c(...)

Arguments

...
A set of objects of class cross. See read.cross for details. These must all either be of the same cross type or be a combination of backcrosses and intercrosses. All crosses must have the same number of chromosomes and chromosome names, and the same marker orders and positions, though the set of markers need not be precisely the same.

Value

The concatenated input, as a cross object. Additional columns are added to the phenotype data indicating which cross an individual comes from; another column indicates cross type (0=BC, 1=intercross), if there are crosses of different types. The crosses are not required to have exactly the same set of phenotypes; phenotypes with the same names are assumed to be the same.

If the crosses have different sets of markers, we interpolate marker order, but the cM positions of markers that are in common between crosses must be precisely the same in the different crosses.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

subset.cross

Examples

data(fake.f2)
junk <- fake.f2
junk <- c(fake.f2,junk)

[Package qtl version 1.74 Index]