c.scanone {qtl}	R Documentation

Combine columns from multiple scanone results

Description

Concatenate the columns from different runs of scanone.

Usage

## S3 method for class 'scanone'
c(..., labels)
## S3 method for class 'scanone'
cbind(..., labels)

Arguments

...
A set of objects of class scanone. (This can also be a list of scanone objects.) These are the results from scanone (with n.perm=0), generally run with different phenotypes or methods. All must conform with each other, meaning that calc.genoprob and/or sim.geno were run with the same values for step and off.end and with data having the same genetic map.

labels
A vector of character strings, of length 1 or of the same length as the input, to be appended to the column names in the output.

Details

The aim of this function is to concatenate the results from multiple runs scanone, generally for different phenotypes and/or methods, to be used in parallel with summary.scanone.

Value

The concatenated input, as a scanone object.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scanone, scanone, cbind.scanoneperm

Examples

data(fake.f2)
fake.f2 <- calc.genoprob(fake.f2)

out.hk <- scanone(fake.f2, method="hk")
out.np <- scanone(fake.f2, model="np")

out <- c(out.hk, out.np, labels=c("hk","np"))
plot(out, lod=1:2, col=c("blue", "red"))

[Package qtl version 1.74 Index]