c.scanoneperm {qtl}	R Documentation

Combine data from scanone permutations

Description

Concatenate the data for multiple runs of scanone with n.perm > 0.

Usage

## S3 method for class 'scanoneperm'
c(...)
## S3 method for class 'scanoneperm'
rbind(...)

Arguments

...
A set of objects of class scanoneperm. (This can also be a list of scanoneperm objects.) These are the permutation results from scanone (that is, when n.perm > 0). These must all have the same number of columns. (That is, they must have been created with the same number of phenotypes, and it is assumed that they were generated in precisely the same way.)

Details

The aim of this function is to concatenate the results from multiple runs of a permutation test scanone, to assist with the case that such permutations are done on multiple processors in parallel.

Value

The concatenated input, as a scanoneperm object.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scanoneperm, scanone, cbind.scanoneperm, c.scantwoperm

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2)
operm1 <- scanone(fake.f2, method="hk", n.perm=100, perm.Xsp=TRUE)
operm2 <- scanone(fake.f2, method="hk", n.perm=50, perm.Xsp=TRUE)

operm <- c(operm1, operm2)

[Package qtl version 1.74 Index]