c.scantwo {qtl}	R Documentation

Combine columns from multiple scantwo results

Description

Concatenate the columns from different runs of scantwo.

Usage

## S3 method for class 'scantwo'
c(...)
## S3 method for class 'scantwo'
cbind(...)

Arguments

...
A set of objects of class scantwo. (This can also be a list of scantwo objects.) These are the results from scantwo (with n.perm=0), generally run with different phenotypes or methods. All must conform with each other, meaning that calc.genoprob and/or sim.geno were run with the same values for step and off.end and with data having the same genetic map.

Details

The aim of this function is to concatenate the results from multiple runs scantwo, generally for different phenotypes and/or methods.

Value

The concatenated input, as a scantwo object.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scantwo, scantwo, c.scanone

Examples

data(fake.bc)
fake.bc <- calc.genoprob(fake.bc)

out2a <- scantwo(fake.bc, method="hk")
out2b <- scantwo(fake.bc, pheno.col=2, method="hk")

out2 <- c(out2a, out2b)

[Package qtl version 1.74 Index]