c.scantwoperm {qtl}	R Documentation

Combine data from scantwo permutations

Description

Concatenate the data for multiple runs of scantwo with n.perm > 0.

Usage

## S3 method for class 'scantwoperm'
c(...)
## S3 method for class 'scantwoperm'
rbind(...)

Arguments

...
A set of objects of class scantwoperm. (This can also be a list of scantwoperm objects.) These are the permutation results from scantwo (that is, when n.perm > 0). These must all concern the same number of LOD columns. (That is, they must have been created with the same number of phenotypes, and it is assumed that they were generated in precisely the same way.)

Details

The aim of this function is to concatenate the results from multiple runs of a permutation test scantwo, to assist with the case that such permutations are done on multiple processors in parallel.

Value

The concatenated input, as a scantwoperm object.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scantwoperm, scantwo, cbind.scantwoperm

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2)
## Not run: operm1 <- scantwo(fake.f2, method="hk", n.perm=50)
operm2 <- scantwo(fake.f2, method="hk", n.perm=50)
## End(Not run)

operm <- c(operm1, operm2)

[Package qtl version 1.74 Index]