calc.errorlod {qtl}	R Documentation

Identify likely genotyping errors

Description

Calculates a LOD score for each genotype, measuring the evidence for genotyping errors.

Usage

calc.errorlod(cross, error.prob=0.01,
              map.function=c("haldane","kosambi","c-f","morgan"),
              version=c("new","old"))

Arguments

cross
An object of class cross. See read.cross for details.

error.prob
Assumed genotyping error rate used in the calculation of the penetrance Pr(observed genotype | true genotype)

map.function
Indicates whether to use the Haldane, Kosambi, Carter-Falconer, or Morgan map function when converting genetic distances into recombination fractions.

version
Specifies whether to use the original version of this function or the current (preferred) version.

Details

Calculates, for each individual at each marker, a LOD score measuring the strength of evidence for a genotyping error, as described by Lincoln and Lander (1992).

In the latest version, evidence for a genotype being in error is considered assuming that all other genotypes (for that individual, on that chromosome) are correct. The argument version allows one to specify whether this new version is used, or whether the original (old) version of the calculation is performed.

Note that values below 4 are generally not interesting. Also note that if markers are extremely tightly linked, recombination events can give large error LOD scores. The error LOD scores should not be trusted blindly, but should be viewed as a tool for identifying genotypes deserving further study.

Use top.errorlod to print all genotypes with error LOD scores above a specified threshold, plotErrorlod to plot the error LOD scores for specified chromosomes, and plotGeno to view the observed genotype data with likely errors flagged.

Value

The input cross object is returned with a component, errorlod, added to each component of cross$geno. The errorlod component is a matrix of size (n.ind x n.mar). An attribute "error.prob" is set to the value of the corresponding argument, for later reference.

Author(s)

Karl W Broman, broman@wisc.edu

References

Lincoln, S. E. and Lander, E. S. (1992) Systematic detection of errors in genetic linkage data. Genomics 14, 604–610.

See Also

plotErrorlod, top.errorlod, cleanGeno

Examples

data(hyper)

hyper <- calc.errorlod(hyper,error.prob=0.01)

# print those above a specified cutoff
top.errorlod(hyper, cutoff=4)

# plot genotype data, flagging genotypes with error LOD > cutoff
plotGeno(hyper, chr=1, ind=160:200, cutoff=7, min.sep=2)

[Package qtl version 1.74 Index]