calc.penalties {qtl}	R Documentation

Calculate LOD penalties

Description

Derive penalties for the penalized LOD scores (used by stepwiseqtl) on the basis of permutation results from a two-dimensional, two-QTL scan (obtained by scantwo).

Usage

calc.penalties(perms, alpha=0.05, lodcolumn)

Arguments

perms
Permutation results from scantwo.

alpha
Significance level.

lodcolumn
If the scantwo permutation results contain LOD scores for multiple phenotypes, this argument indicates which to use in the summary. This may be a vector. If missing, penalties for all phenotypes are calculated.

Details

Thresholds derived from scantwo permutations (that is, for a two-dimensional, two-QTL genome scan) are used to calculate penalties on main effects and interactions.

The main effect penalty is the 1-alpha quantile of the null distribution of the genome-wide maximum LOD score from a single-QTL genome scan (as with scanone).

The "heavy" interaction penalty is the 1-alpha quantile of the null distribution of the maximum interaction LOD score (that is, the log10 likelihood ratio comparing the best model with two interacting QTL to the best model with two additive QTL) from a two-dimensional, two-QTL genome scan (as with scantwo).

The "light" interaction penality is the difference between the "fv1" threshold from the scantwo permutations (that is, the 1-alpha quantile of the LOD score comparing the best model with two interacting QTL to the best single-QTL model) and the main effect penalty.

If the permutations results were obtained with perm.Xsp=TRUE, to give X-chr-specific results, six penalties are calculated: main effect for autosomes, main effect for X chr, heavy penalty on A:A interactions, light penalty on A:A interactions, penalty on A:X interactions, and penalty on X:X interactions.

Value

Vector of three values indicating the penalty on main effects and heavy and light penalties on interactions, or a matrix of such results, with each row corresponding to a different phenotype.

If the input permutations are X-chromosome-specific, the result has six values: main effect for autosomes, main effect for X chr, heavy penalty on A:A interactions, light penalty on A:A interactions, penalty on A:X interactions, and penalty on X:X interactions.

Author(s)

Karl W Broman, broman@wisc.edu

References

Manichaikul, A., Moon, J. Y., Sen, Ś, Yandell, B. S. and Broman, K. W. (2009) A model selection approach for the identification of quantitative trait loci in experimental crosses, allowing epistasis. Genetics, 181, 1077–1086.

See Also

scantwo, stepwiseqtl

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2, step=5)
out.2dim <- scantwo(fake.f2, method="hk")

# permutations

## Not run: permo.2dim <- scantwo(fake.f2, method="hk", n.perm=1000)
summary(permo.2dim, alpha=0.05)

# penalties
calc.penalties(permo.2dim)

[Package qtl version 1.74 Index]