cbind.scanoneperm {qtl}	R Documentation

Combine columns from multiple scanone permutation results

Description

Concatenate the columns from different runs of scanone with n.perm > 0.

Usage

## S3 method for class 'scanoneperm'
cbind(..., labels)

Arguments

...
A set of objects of class scanoneperm. These are the permutation results from scanone (that is, when n.perm > 0), generally run with different phenotypes or methods.

labels
A vector of character strings, of length 1 or of the same length as the input ..., to be appended to the column names in the output.

Details

The aim of this function is to concatenate the results from multiple runs of a permutation test scanone, generally for different phenotypes and/or methods, to be used in parallel with c.scanone.

Value

The concatenated input, as a scanoneperm object. If different numbers of permutation replicates were used, those columns with fewer replicates are padded with missing values (NA).

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scanoneperm, scanone, c.scanoneperm, c.scanone

Examples

data(fake.f2)
fake.f2 <- calc.genoprob(fake.f2)

operm1 <- scanone(fake.f2, method="hk", n.perm=10, perm.Xsp=TRUE)
operm2 <- scanone(fake.f2, method="em", n.perm=5, perm.Xsp=TRUE)

operm <- cbind(operm1, operm2, labels=c("hk","em"))
summary(operm)

[Package qtl version 1.74 Index]