cbind.scantwoperm {qtl}	R Documentation

Combine scantwo permutations by column

Description

Column-bind permutations results from scantwo for multiple phenotypes or models.

Usage

## S3 method for class 'scantwoperm'
cbind(...)

Arguments

...
A set of objects of class scantwoperm. (This can also be a list of scantwoperm objects.) These are the permutation results from scantwo (that is, when n.perm > 0). These must all concern the same number of permutations.

Value

The column-binded input, as a scantwoperm object.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scantwo, c.scantwoperm, summary.scantwoperm

Examples

data(fake.bc)

fake.bc <- calc.genoprob(fake.bc)
## Not run: operm1 <- scantwo(fake.bc, pheno.col=1, method="hk", n.perm=50)
operm2 <- scantwo(fake.bc, pheno.col=2, method="hk", n.perm=50)
## End(Not run)

operm <- cbind(operm1, operm2)

[Package qtl version 1.74 Index]