checkAlleles {qtl}	R Documentation

Identify markers with switched alleles

Description

Identify markers whose alleles might have been switched by comparing the LOD score for linkage to all other autosomal markers with the original data to that when the alleles have been switched.

Usage

checkAlleles(cross, threshold=3, verbose)

Arguments

cross
An object of class cross. See read.cross for details.

threshold
Only an increase in maximum 2-point LOD of at least this amount will lead to a marker being flagged.

verbose
If TRUE and there are no markers above the threshold, print a message.

Details

For each marker, we compare the maximum LOD score for the cases where the estimated recombination fraction > 0.5 to those where r.f. < 0.5. The function est.rf must first be run.

Note: Markers that are tightly linked to a marker whose alleles are switched are likely to also be flagged by this method. The real problem markers are likely those with the biggest difference in LOD scores.

Value

A data frame containing the flagged markers, having four columns: the marker name, chromosome ID, numeric index within chromosome, and the difference between the maximum two-point LOD score with the alleles switched to that from the original data.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

est.rf, geno.crosstab, switchAlleles

Examples

data(fake.f2)

# switch homozygotes at marker D5M391
fake.f2 <- switchAlleles(fake.f2, "D5M391")

fake.f2 <- est.rf(fake.f2)
checkAlleles(fake.f2)

[Package qtl version 1.74 Index]