chrlen {qtl}	R Documentation

Chromosome lengths in QTL experiment

Description

Obtain the chromosome lengths in a cross or map object.

Usage

chrlen(object)

Arguments

object
An object of class map or of class cross.

Value

Returns a vector of chromosome lengths. If the cross has sex-specific maps, it returns a 2-row matrix with the two lengths for each chromosome.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summaryMap, pull.map, summary.cross

Examples

data(fake.f2)
chrlen(fake.f2)

map <- pull.map(fake.f2)
chrlen(map)

[Package qtl version 1.74 Index]