clean.scantwo {qtl}	R Documentation

Clean up scantwo output

Description

In an object output from scantwo, replaces negative and missing LOD scores with 0, and replaces LOD scores for pairs of positions that are not separated by n.mar markers, or that are less than distance cM apart, with 0. Further, if the LOD for full model is less than the LOD for the additive model, the additive LOD is pasted over the full LOD.

Usage

## S3 method for class 'scantwo'
clean(object, n.mar=1, distance=0, ...)

Arguments

object
An object of class scantwo. See scantwo for details.

n.mar
Pairs of positions not separated by at least this many markers have LOD scores set to 0.

distance
Pairs of positions not separated by at least this distance have LOD scores set to 0.

...
Ignored at this point.

Value

The input scantwo object, with any negative or missing LOD scores replaced by 0, and LOD scores for pairs of positions separated by fewer than n.mar markers, or less than distance cM, are set to 0. Also, if the LOD for the full model is less than the LOD for the additive model, the additive LOD is used in place of the full LOD.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scantwo, summary.scantwo

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2, step=5)
out2 <- scantwo(fake.f2, method="hk")
out2 <- clean(out2)
out2cl2 <- clean(out2, n.mar=2, distance=5)

[Package qtl version 1.74 Index]