cleanGeno {qtl}	R Documentation

Delete genotypes that are possibly in error

Description

Delete genotypes from a cross that are indicated to be possibly in error, as they result in apparent tight double-crossovers.

Usage

cleanGeno(cross, chr, maxdist=2.5, maxmark=2, verbose=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

maxdist
A vector specifying the maximum distance between two crossovers.

maxmark
A vector specifying the maximum number of typed markers between two crossovers.

verbose
If TRUE, print information on the numbers of genotypes omitted from each chromosome.

Details

We first use locateXO to identify crossover locations. If a pair of adjacted crossovers are separated by no more than maxdist and contain no more than maxmark genotyped markers, the intervening genotypes are omitted (that is, changed to NA).

The arguments maxdist and maxmark may be vectors. (If both have length greater than 1, they must have the same length.) If they are vectors, genotypes are omitted if they satisify any one of the (maxdist, maxmark) pairs.

Value

The input cross object with suspect genotypes omitted.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

locateXO, countXO, calc.errorlod

Examples

data(hyper)
sum(ntyped(hyper))
hyperc <- cleanGeno(hyper, chr=4, maxdist=c(2.5, 10), maxmark=c(2, 1))
sum(ntyped(hyperc))

[Package qtl version 1.74 Index]