comparecrosses {qtl}	R Documentation

Compare two cross objects

Description

Verify that two objects of class cross have identical classes, chromosomes, markers, genotypes, genetic maps, and phenotypes.

Usage

comparecrosses(cross1, cross2, tol=1e-5)

Arguments

cross1
An object of class cross (must be an intercross). See read.cross for details.

cross2
An object of class cross (must be an intercross). See read.cross for details.

tol
Tolerance value for comparing genetic map positions and numeric phenotypes.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.cross

Examples

data(listeria)
comparecrosses(listeria, listeria)

[Package qtl version 1.74 Index]