comparegeno {qtl}	R Documentation

Compare individuals' genotype data

Description

Count proportion of matching genotypes between all pairs of individuals, to look for unusually closely related individuals.

Usage

comparegeno(cross, what=c("proportion","number","both"))

Arguments

cross
An object of class cross. See read.cross for details.

what
Indicates whether to return the proportion or number of matching genotypes (or both).

Value

A matrix whose (i,j)th element is the proportion or number of matching genotypes for individuals i and j.

If called with what="both", the lower triangle contains the proportion and the upper triangle contains the number.

If called with what="proportion", the diagonal contains missing values. Otherwise, the diagonal contains the number of typed markers for each individual.

The output is given class "comparegeno" so that appropriate summary and plot functions may be used.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

nmissing, summary.comparegeno, plot.comparegeno

Examples

data(listeria)

cg <- comparegeno(listeria)

summary(cg, 0.7)
plot(cg)

[Package qtl version 1.74 Index]