condense.scantwo {qtl}	R Documentation

Condense the output from a 2-d genome scan

Description

Produces a very condensed version of the output of scantwo.

Usage

## S3 method for class 'scantwo'
condense(object)

Arguments

object
An object of class scantwo, the output of the function scantwo.

Details

This produces a very reduced version of the output of scantwo, for which a summary may still be created via summary.scantwo, though plots can no longer be made.

Value

An object of class scantwocondensed, containing just the maximum full, additive and interactive LOD scores, and the positions where they occured, on each pair of chromosomes.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scantwo, summary.scantwo, max.scantwo

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2)

out2 <- scantwo(fake.f2, method="hk")

out2c <- condense(out2)
summary(out2c, allpairs=FALSE)
max(out2c)

[Package qtl version 1.74 Index]