convert.map {qtl}	R Documentation

Change map function for a genetic map

Description

Convert a genetic map from using one map function to another.

Usage

## S3 method for class 'map'
convert(object, old.map.function=c("haldane", "kosambi", "c-f", "morgan"),
         new.map.function=c("haldane", "kosambi", "c-f", "morgan"), ...)

Arguments

object
A genetic map object, of class "map": A list whose components are vectors of marker locations.

old.map.function
The map function used in forming the map in object.

new.map.function
The new map function to be used.

...
Ignored at this point.

Details

The location of the first marker on each chromosome is left unchanged. Inter-marker distances are converted to recombination fractions with the inverse of the old.map.function, and then back to distances with the new.map.function.

Value

The same as the input, but with inter-marker distances changed to reflect a different map function.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

est.map, replace.map

Examples

data(listeria)
map <- pull.map(listeria)
map <- convert(map, "haldane", "kosambi")
listeria <- replace.map(listeria, map)

[Package qtl version 1.74 Index]