convert.scanone {qtl}	R Documentation

Convert output from scanone for R/qtl version 0.98

Description

Convert the output from scanone from the format used in R/qtl version 0.97 and earlier to that used in version 0.98 and later.

Usage

## S3 method for class 'scanone'
convert(object, ...)

Arguments

object
Output from the function scanone, for R/qtl version 0.97 and earlier.

...
Ignored at this point.

Details

Previously, inter-marker locations were named as, for example, loc7.5.c3; these were changed to c3.loc7.5.

Value

The same scanone output, but revised for use with R/qtl version 0.98 and later.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scanone, convert.scantwo

Examples

## Not run: out.new <- convert(out.old)

[Package qtl version 1.74 Index]