convert.scantwo {qtl}	R Documentation

Convert output from scantwo for R/qtl version 1.03 and earlier

Description

Convert the output from scantwo from the format used in R/qtl version 1.03 and earlier to that used in version 1.04 and later.

Usage

## S3 method for class 'scantwo'
convert(object, ...)

Arguments

object
Output from the function scantwo, for R/qtl version 1.03 and earlier.

...
Ignored at this point.

Details

Previously, the output from scantwo contained the full and interaction LOD scores. In R/qtl version 1.04 and later, the output contains the LOD scores from the full and additive QTL models.

Value

The same scanone output, but revised for use with R/qtl version 1.03 and later.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scantwo, convert.scanone

Examples

## Not run: out2.new <- convert(out2.old)

[Package qtl version 1.74 Index]