convert2risib {qtl}	R Documentation

Convert a cross to RIL by sib mating

Description

Convert a cross to type "risib" (RIL by sib mating).

Usage

convert2risib(cross)

Arguments

cross
An object of class cross. See read.cross for details.

Details

If there are more genotypes with code 3 (BB) than code 2 (AB), we omit the genotypes with code==2 and call those with code==3 the BB genotypes.

If, instead, there are more genotypes with code 2 than code 3, we omit the genotypes with code==3 and call those with code==2 the BB genotypes.

Value

The input cross object, with genotype codes possibly changed and cross type changed to "risib".

Author(s)

Karl W Broman, broman@wisc.edu

See Also

convert2riself

Examples

data(hyper)
hyper.as.risib <- convert2risib(hyper)

[Package qtl version 1.74 Index]