convert2sa {qtl}	R Documentation

Convert a sex-specific map to a sex-averaged one

Description

Convert a sex-specific map to a sex-averaged one, assuming that the female and male maps are actually the same (that is, that the map was estimated assuming a common recombination rate in females and males).

Usage

convert2sa(map, tol=1e-4)

Arguments

map
A map object with sex-specific locations (but assuming that the female and male maps are the same), as output by the function est.map for a 4-way cross, with argument sex.sp=FALSE.

tol
Tolerance value for inspecting the differences between the female and male maps; if they differ by more than this tolerance, a warning is issued.

Details

We pull out just the female marker locations, and give a warning if there are large differences between the female and male maps.

Value

A map object, with sex-averaged distances.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

est.map, plotMap

Examples

data(fake.4way)
## Not run: fake.4way <- subset(fake.4way, chr="-X")

nm <- est.map(fake.4way, sex.sp=FALSE)
plot(convert2sa(nm))

[Package qtl version 1.74 Index]