countXO {qtl}	R Documentation

Count number of obligate crossovers for each individual

Description

Count the number of obligate crossovers for each individual in a cross, either by chromosome or overall.

Usage

countXO(cross, chr, bychr=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to investigate. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

bychr
If TRUE, return counts for each individual chromosome; if FALSE, return the overall number across the selected chromosomes.

Details

For each individual we count the minimal number of crossovers that explain the observed genotype data.

Value

If bychr=TRUE, a matrix of counts is returned, with rows corresponding to individuals and columns corresponding to chromosomes.

If bychr=FALSE, a vector of counts (the total number of crossovers across all selected chromosomes) is returned.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

ripple, locateXO, cleanGeno

Examples

data(hyper)
plot(countXO(hyper))

[Package qtl version 1.74 Index]