drop.dupmarkers {qtl}	R Documentation

Drop duplicate markers

Description

Drop markers with duplicate names; retaining the first of each set, with consensus genotyps

Usage

drop.dupmarkers(cross, verbose=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

verbose
If TRUE, print information on the numbers of genotypes and markers omitted. If > 1, give more detailed information on genotypes omitted.

Value

The input cross object, with any duplicate markers omitted (except for one). The marker retained will have consensus genotypes; if multiple versions of a marker have different genotypes for an individual, they will be replaced by NA.

Any derived data (such as produced by calc.genoprob) will be stripped off.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

drop.nullmarkers, pull.markers, drop.markers, summary.cross, clean.cross

Examples

data(listeria)

listeria <- drop.dupmarkers(listeria)

[Package qtl version 1.74 Index]