drop.markers {qtl}	R Documentation

Drop a set of markers

Description

Drop a vector of markers from the data matrices and genetic maps.

Usage

drop.markers(cross, markers)

Arguments

cross
An object of class cross. See read.cross for details.

markers
A character vector of marker names.

Value

The input object, with any markers in the vector markers removed from the genotype data matrices, genetic maps, and, if applicable, any derived data (such as produced by calc.genoprob). (It might be a good idea to re-derive such things after using this function.)

Author(s)

Karl W Broman, broman@wisc.edu

See Also

drop.nullmarkers, pull.markers, geno.table, clean.cross

Examples

data(listeria)
listeria2 <- drop.markers(listeria, c("D10M44","D1M3","D1M75"))

[Package qtl version 1.74 Index]