drop.nullmarkers {qtl}	R Documentation

Drop markers without any genotype data

Description

Drop markers, from the data matrices and genetic maps, that have no genotype data.

Usage

drop.nullmarkers(cross)

Arguments

cross
An object of class cross. See read.cross for details.

Value

The input object, with any markers lacking genotype data removed from the genotype data matrices, genetic maps, and, if applicable, any derived data (such as produced by calc.genoprob). (It might be a good idea to re-derive such things after using this function.)

Author(s)

Karl W Broman, broman@wisc.edu

See Also

nullmarkers, drop.markers, clean.cross, geno.table

Examples

# removes one marker from hyper
data(hyper)
hyper <- drop.nullmarkers(hyper)

# shouldn't do anything to listeria
data(listeria)
listeria <- drop.nullmarkers(listeria)

[Package qtl version 1.74 Index]