est.rf {qtl}	R Documentation

Estimate pairwise recombination fractions

Description

Estimate the sex-averaged recombination fraction between all pairs of genetic markers.

Usage

est.rf(cross, maxit=10000, tol=1e-6)

Arguments

cross
An object of class cross. See read.cross for details.

maxit
Maximum number of iterations for the EM algorithm (not used with backcrosses).

tol
Tolerance for determining convergence (not used with backcrosses).

Details

For a backcross, one can simply count recombination events. For an intercross or 4-way cross, a version of the EM algorithm must be used to estimate recombination fractions. (Since, for example, in an intercross individual that is heterozygous at two loci, it is not known whether there were 0 or 2 recombination events.) Note that, for the 4-way cross, we estimate sex-averaged recombination fractions.

Value

The input cross object is returned with a component, rf, added. This is a matrix of size (tot.mar x tot.mar). The diagonal contains the number of typed meioses per marker, the lower triangle contains the estimated recombination fractions, and the upper triangle contains the LOD scores (testing rf = 0.5).

Author(s)

Karl W Broman, broman@wisc.edu

See Also

plotRF, pull.rf, plot.rfmatrix, est.map, badorder, checkAlleles

Examples

data(badorder)
badorder <- est.rf(badorder)
plotRF(badorder)

[Package qtl version 1.74 Index]