fake.f2 {qtl}	R Documentation

Simulated data for an F2 intercross

Description

Simulated data for an F2 intercross, obtained using sim.cross.

Usage

data(fake.f2)

Format

An object of class cross. See read.cross for details.

Details

There are 200 F2 individuals typed at 94 markers, including 3 on the X chromosome. There is one quantitative phenotype, along with an indication of sex (0=female, 1=male) and the direction of the cross (pgm = paternal grandmother, 0=A, meaning the cross was (AxB)x(AxB), and 1=B, meaning the cross was (AxB)x(BxA)).

Note that the X chromosome genotypes are coded in a special way (see read.cross). For the individuals with pgm=0, sex=0, 1=AA and 2=AB; for individuals with pgm=0, sex=1, 1=A and 2=B (hemizygous); for individuals with pgm=1, sex=0, 1=BB and 2=AB; for individuals with pgm=1, sex=1, 1=A and 2=B. This requires special care!

The data were simulated using an additive model with three QTLs on chromosome 1 (at 30, 50 and 70 cM), one QTL on chromosome 13 (at 30 cM), and one QTL on the X chromosome (at 10 cM).

Author(s)

Karl W Broman, broman@wisc.edu

See Also

sim.cross, fake.bc, fake.4way, listeria, hyper, bristle3, bristleX

Examples

data(fake.f2)
summary(fake.f2)
plot(fake.f2)

[Package qtl version 1.74 Index]