find.flanking {qtl}	R Documentation

Find flanking markers for a specified position

Description

Find the genetic markers flanking a specified position on a chromosome, as well as the marker that is closest to the specified position.

Usage

find.flanking(cross, chr, pos)

Arguments

cross
An object of class cross. See read.cross for details.

chr
A vector of chromosome identifiers, or a single such.

pos
A vector of cM positions.

Value

A data.frame, each row corresponding to one of the input positions. The first column contains the left-flanking markers, the second column contains the right-flanking markers, and the third column contains the markers closest to the specified positions.

Author(s)

Brian Yandell

See Also

find.marker, plotPXG, find.markerpos, find.pseudomarker

Examples

data(listeria)
find.flanking(listeria, 5, 28)
find.flanking(listeria, c(1, 5, 13), c(81, 28, 26))

[Package qtl version 1.74 Index]