find.marker {qtl}	R Documentation

Find marker closest to a specified position

Description

Find the genetic marker closest to a specified position on a chromosome.

Usage

find.marker(cross, chr, pos, index)

Arguments

cross
An object of class cross. See read.cross for details.

chr
A vector of chromosome identifiers, or a single such.

pos
A vector of cM positions.

index
A vector of numeric indices of the markers within chromosomes.

Details

Provide one of pos or index.

If the input chr has length one, it is expanded to the same length as the input pos or index.

If pos is specified and multiple markers are exactly the same distance from the specified position, one is chosen at random from among those with the most genotype data.

For a cross with sex-specific maps, positions specified by pos are assumed to correspond to the female genetic map.

Value

A vector of marker names (of the same length as the input pos), corresponding to the markers nearest to the specified chromosomes/positions (if pos is specified) or to the input numeric indices (in index is specified).

Author(s)

Karl W Broman, broman@wisc.edu

See Also

find.flanking, plotPXG, find.pseudomarker, effectplot, find.markerpos

Examples

data(listeria)
find.marker(listeria, 5, 28)
find.marker(listeria, 5, index=6)
find.marker(listeria, c(1, 5, 13), c(81, 28, 26))

[Package qtl version 1.74 Index]