find.pheno {qtl}	R Documentation

Find column number for a particular phenotype

Description

Find the column number corresponding to a particular phenotype name.

Usage

find.pheno(cross, pheno)

Arguments

cross
An object of class cross. See read.cross for details.

pheno
Vector of phenotype names (as character strings).

Value

A vector of numbers, corresponding to the column numbers of the phenotype in the input cross with the specified names.

Author(s)

Brian Yandell

Examples

data(fake.bc)
find.pheno(fake.bc, "sex")

[Package qtl version 1.74 Index]