find.pseudomarker {qtl}	R Documentation

Find the pseudomarker closest to a specified position

Description

Find the pseudomarker closest to a specified position on a chromosome.

Usage

find.pseudomarker(cross, chr, pos, where=c("draws", "prob"), addchr=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
A vector of chromosome identifiers, or a single such.

pos
A vector of cM positions.

where
Indicates whether to look in the draws or prob components of the input cross.

addchr
If TRUE, include something like "c5." at the beginning of the names of non-pseudomarker locations, as in the output of scanone; if FALSE, don't include this sort of string, as in the genotype probabilities from calc.genoprob.

Details

If the input chr has length one, it is expanded to the same length as the input pos.

If multiple markers are exactly the same distance from the specified position, one is chosen at random from among those with the most genotype data.

For a cross with sex-specific maps, the input positions are assumed to correspond to the female genetic map.

Value

A vector of pseudomarker names (of the same length as the input pos), corresponding to the markers nearest to the specified chromosomes/positions.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

find.flanking, plotPXG, effectplot, find.marker, find.markerpos

Examples

data(listeria)
listeria <- calc.genoprob(listeria, step=2.5)
find.pseudomarker(listeria, 5, 28, "prob")
find.pseudomarker(listeria, c(1, 5, 13), c(81, 28, 26), "prob")

[Package qtl version 1.74 Index]