find_large_intervals {qtl}	R Documentation

Find large intervals in a map

Description

Find large inter-marker intervals in a map.

Usage

find_large_intervals(map, min_length=35)

Arguments

map
A list of numeric vectors; each component is a chromosome with the positions of markers on that chromosome. Can also be an object of class cross, in which case pull.map is used.

min_length
Minimum length of interval to be flagged.

Value

Data frame with chromosome, left and right markers and interval length.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summaryMap

Examples

data(fake.f2)

find_large_intervals(fake.f2, 30)

[Package qtl version 1.74 Index]