find.markerindex {qtl}	R Documentation

Determine the numeric index for a marker

Description

Determine the numeric index for a marker in a cross object, when all markers on all chromosomes are pasted together.

Usage

find.markerindex(cross, name)

Arguments

cross
An object of class cross. See read.cross for details.

name
A vector of marker names.

Value

A vector of numeric indices, from 1, 2, ..., totmar(cross), with NA for markers not found.

Author(s)

Danny Arends; Karl W Broman broman@wisc.edu

See Also

find.markerpos

Examples

data(hyper)
mar <- find.marker(hyper, 4, 30)
find.markerindex(hyper, mar)

[Package qtl version 1.74 Index]