flip.order {qtl}	R Documentation

Flip the orders of markers on a set of chromosomes

Description

Flip the orders of markers on a specified set of chromosome, so that the markers will be in the reverse order.

Usage

flip.order(cross, chr)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Vector indicating the chromosomes to flip. This should be a vector of character strings referring to chromosomes by name. A logical (TRUE/FALSE) vector may also be used.

Details

If the cross contains results from calc.genoprob, sim.geno, argmax.geno, or calc.errorlod, those results are also updated.

Results of est.rf and markerlrt are deleted.

Value

The input cross object, but with the marker order on the specified chromosomes flipped.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

switch.order

Examples

data(fake.f2)
fake.f2 <- flip.order(fake.f2, c(1, 5, 13))

[Package qtl version 1.74 Index]