formLinkageGroups {qtl}	R Documentation

Partition markers into linkage groups

Description

Use pairwise linkage information between markers (as calculated by est.rf to partition markers into linkage groups.

Usage

formLinkageGroups(cross, max.rf=0.25, min.lod=3, reorgMarkers=FALSE,
                  verbose=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

max.rf
Maximum recombination fraction for placing two markers in the same linkage group (see Details).

min.lod
Minimum LOD score for placing two markers in the same linkage group (see Details).

reorgMarkers
If TRUE, the output is a cross object, like the input, but with the markers organized into the inferred linkage groups. If FALSE, the output is a table indicating the initial chromosome assignments and the inferred linkage group partitions.

verbose
If TRUE, display information about the progress of the calculations.

Details

Two markers are placed in the same linkage group if the estimated recombination fraction between them is <= max.rf and the LOD score (for the test of the rec. frac. = 1/2) is >= min.lod. The transitive property (if A is linked to B and B is linked to C then A is linked to C) is used to close the groups.

Value

If reorgMarkers=FALSE (the default), the output is a data frame with rows corresponding to the markers and with two columns: the initial chromosome assignment and the inferred linkage group. Linkage groups are ordered by the number of markers they contain (from largest to smallest).

If reorgMarkers=TRUE, the output is a cross object, like the input, but with the markers reorganized into the inferred linkage groups. The marker order and marker positions within the linkage groups are arbitrary.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

est.rf, orderMarkers

Examples

data(listeria)
listeria <- est.rf(listeria)
result <- formLinkageGroups(listeria)
tab <- table(result[,1], result[,2])
apply(tab, 1, function(a) sum(a!=0))
apply(tab, 2, function(a) sum(a!=0))

[Package qtl version 1.74 Index]