formMarkerCovar {qtl}	R Documentation

Create matrix of marker covariates for QTL analysis

Description

Pull out a matrix of genotypes or genotype probabilities to use markers as covariates in QTL analysis.

Usage

formMarkerCovar(cross, markers, method=c("prob", "imp", "argmax"), ...)

Arguments

cross
An object of class cross. See read.cross for details.

markers
A vector of character strings of marker or pseudomarker names. Pseudomarker names may be of the form "5@21.5" (for chr 5 at 21.5 cM), but then all names must be of this form.

method
If method="prob", the genotype probabilities from calc.genoprob are used; otherwise we use fill.geno to impute missing data, with this method.

...
Passed to fill.geno, if necessary.

Value

A matrix containing genotype probabilities or genotype indicators, suitable for use as covariates in scanone.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.geno, pull.genoprob, fill.geno, scanone

Examples

data(hyper)
hyper <- calc.genoprob(hyper, step=0)
peakMarker <- "D4Mit164"
X <- formMarkerCovar(hyper, peakMarker)

out <- scanone(hyper, addcovar=X)

[Package qtl version 1.74 Index]