geno.crosstab {qtl}	R Documentation

Create table of two-locus genotypes

Description

Create a cross tabulation of the genotypes at a pair of markers.

Usage

geno.crosstab(cross, mname1, mname2, eliminate.zeros=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

mname1
The name of the first marker (as a character string). (Alternatively, a vector with the two character strings, in which case mname2 should not be given.)

mname2
The name of the second marker (as a character string).

eliminate.zeros
If TRUE, don't show the rows and columns that have no data.

Value

A matrix containing the number of individuals having each possible pair of genotypes. Genotypes for the first marker are in the rows; genotypes for the second marker are in the columns.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

geno.table, find.marker

Examples

data(hyper)
geno.crosstab(hyper, "D1Mit123", "D1Mit156")
geno.crosstab(hyper, "DXMit22", "DXMit16")
geno.crosstab(hyper, c("DXMit22", "DXMit16"))

[Package qtl version 1.74 Index]