geno.image {qtl}	R Documentation

Plot grid of genotype data

Description

Plot a grid showing which the genotype data in a cross.

Usage

geno.image(x, chr, reorder=FALSE, main="Genotype data",
           alternate.chrid=FALSE, col=NULL, ...)

Arguments

x
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to be drawn in the plot. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

reorder
Specify whether to reorder individuals according to their phenotypes.

FALSE 	 Don't reorder

TRUE 	 Reorder according to the sum of the phenotypes

n 	 Reorder according to phenotype n

main
Title to place on plot.

alternate.chrid
If TRUE and more than one chromosome is plotted, alternate the placement of chromosome axis labels, so that they may be more easily distinguished.

col
Vector of colors. The first is for missing genotypes, followed by colors for each of the genotypes. If NULL, a default set of colors are used.

...
Passed to image.

Details

Uses image to plot a grid with the genotype data. The genotypes AA, AB, BB are displayed in the colors red, blue, and green, respectively. In an intercross, if there are genotypes "not BB" and "not AA", these are displayed in purple and orange, respectively. White pixels indicate missing data.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

plot.cross, plotMissing, plotGeno, image

Examples

data(listeria)
geno.image(listeria)

[Package qtl version 1.74 Index]