getid {qtl}	R Documentation

Pull out the individual identifiers from a cross

Description

Pull out the individual identifiers from a cross object.

Usage

getid(cross)

Arguments

cross
An object of class cross. See read.cross for details.

Value

A vector of individual identifiers, pulled from the phenotype data (a column named id or ID).

If there are no such identifiers in the cross, the function returns NULL.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

subset.cross, top.errorlod

Examples

data(fake.f2)

# create an ID column
fake.f2$pheno$id <- paste("ind", sample(nind(fake.f2)), sep="")

getid(fake.f2)

[Package qtl version 1.74 Index]