groupclusteredheatmap {qtl}	R Documentation

Retrieving groups of traits after clustering

Description

Retrieving groups of clustered traits from the output of mqmplot.clusteredheatmap.

Usage

  groupclusteredheatmap(cross, clusteredheatmapresult, height)

Arguments

cross
An object of class cross. See read.cross for details.

clusteredheatmapresult
Resultint dendrogram object from mqmplot.clusteredheatmap

height
Height at which to 'cut' the dendrogram, a higher cut-off gives less but larger groups. Height represents the maximum distance between two traits clustered together using hclust. the 'normal' behaviour of bigger groups when using a higher heigh cut-off depends on the tree stucture and the amount of traits clustered using mqmplot.clusteredheatmap

Value

A list containing groups of traits which were clustered together with a distance less that height

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

multitrait <- fill.geno(multitrait) # impute missing genotype data
result <- mqmscanall(multitrait, logtransform=TRUE)
cresults <- mqmplot.clusteredheatmap(multitrait,result)
groupclusteredheatmap(multitrait,cresults,10)

[Package qtl version 1.74 Index]