hyper {qtl}	R Documentation

Data on hypertension

Description

Data from an experiment on hypertension in the mouse.

Usage

data(hyper)

Format

An object of class cross. See read.cross for details.

Details

There are 250 male backcross individuals typed at 174 markers (actually one contains only missing values), including 4 on the X chromosome, with one phenotype.

The phenotype is the blood pressure. See the reference below. Note that, for most markers, genotypes are available on only the individuals with extreme phenotypes. At many markers, only recombinant individuals were typed.

Source

Bev Paigen and Gary Churchill (The Jackson Laboratory, Bar Harbor, Maine) https://phenome.jax.org/projects/Sugiyama2

References

Sugiyama, F., Churchill, G. A., Higgens, D. C., Johns, C., Makaritsis, K. P., Gavras, H. and Paigen, B. (2001) Concordance of murine quantitative trait loci for salt-induced hypertension with rat and human loci. Genomics 71, 70–77.

See Also

fake.bc, fake.f2, fake.4way, listeria, bristle3, bristleX

Examples

data(hyper)
summary(hyper)
plot(hyper)

# Note the selective genotyping
## Not run: plotMissing(hyper, reorder=TRUE)

# A marker on c14 has no data; remove it
hyper <- drop.nullmarkers(hyper)

[Package qtl version 1.74 Index]