inferFounderHap {qtl}	R Documentation

Crude reconstruction of founder haplotypes in multi-parent RIL

Description

Uses groups of adjacent markers to infer the founder haplotypes in SNP data on multi-parent recombinant inbred lines.

Usage

inferFounderHap(cross, chr, max.n.markers=15)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Indicator of chromosome to consider. If multiple chromosomes are selected, only the first is used.

max.n.markers
Maximum number of adjacent markers to consider.

Details

We omit SNPs for which any of the founders are missing.

We then consider groups of adjacent SNPs, looking for founder haplotypes that are unique; RIL sharing such a unique haplotype are then inferred to have that founder's DNA.

We consider each marker as the center of a haplotype, and consider haplotypes of size 1, 3, 5, ..., max.n.markers. We end the extension of the haplotypes when all founders have a unique haplotype.

Value

A matrix of dimension nind(cross) x no. markers, with the inferred founder origin for each line at each marker.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

sim.geno, calc.genoprob, fill.geno, argmax.geno

Examples

map <- sim.map(100, n.mar=101, include.x=FALSE, eq.spacing=TRUE)
founderGeno <- simFounderSnps(map, "8")
ril <- sim.cross(map, n.ind=10, type="ri8sib", founderGeno=founderGeno)

h <- inferFounderHap(ril, max.n.markers=11)
mean(!is.na(h)) # proportion inferred
plot(map[[1]], h[1,], ylim=c(0.5, 8.5), xlab="Position", ylab="Genotype")

[Package qtl version 1.74 Index]