jittermap {qtl}	R Documentation

Jitter marker positions in a genetic map

Description

Jitter the marker positions in a genetic map so that no two markers are on top of each other.

Usage

jittermap(object, amount=1e-6)

Arguments

object
Either a cross (an object of class cross; see read.cross for details) or a map (an object of class map; see pull.map for details).

amount
The amount by which markers should be moved.

Value

Either the input cross object or the input map, but with marker positions slightly jittered. If the input was a cross, the function clean is run to strip off any intermediate calculations.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.map, replace.map, summary.cross

Examples

data(hyper)
hyper <- jittermap(hyper)

[Package qtl version 1.74 Index]