locateXO {qtl}	R Documentation

Estimate locations of crossovers

Description

Estimate the locations of crossovers for each individual on a given chromosome.

Usage

locateXO(cross, chr, full.info=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Chromosome to investigate (if unspecified, the first chromosome is considered). This should be a character string referring to a chromosome by name; numeric values are converted to strings.

full.info
If TRUE, output will include information on the left and right endpoints of the intervals to which recombination events are known, as well as the corresponding marker indices.

Details

For each individual we detemine the locations of obligate crossovers, and estimate their location to be at the midpoint between the nearest flanking typed markers.

The function currently only works for a backcross, intercross, or recombinant inbred line.

Value

A list with one component per individual. Each component is either NULL or is a numeric vector with the estimated crossover locations.

If full.info=TRUE, in place of a numeric vector with estimated locations, there is a matrix that includes those locations, the left and right endpoints of the intervals to which crossovers can be placed, the marker indices corresponding to those endpoint, and genotype codes for the genotypes to the left and right of each crossover. The final column indicates the number of typed markers between the current crossover and the next one (useful for identifying potential genotyping errors).

Author(s)

Karl W Broman, broman@wisc.edu

See Also

countXO, cleanGeno

Examples

data(hyper)
xoloc <- locateXO(hyper, chr=4)
table(sapply(xoloc, length))

[Package qtl version 1.74 Index]