locations {qtl}	R Documentation

Genetic locations of traits for the multitrait dataset

Description

A table with genetic locations of the traits in the multitrait dataset

Usage

data(locations)

Format

Each row is a trait with the following information: Name, Name of the trait (will be checked against the name in the cross object Chr, Chromosome of the trait cM, Location in cM from the start of the chromosome

Source

Additional information from the Arabidopsis RIL selfing experiment with Landsberg erecta (Ler) and Cape Verde Islands (Cvi) with 162 individuals scored (with errors at) 117 markers. Dataset obtained from GBIC - Groningen BioInformatics Centre

References

Keurentijes JJB, Fu J, de Vos CHR,Lommen A, Jansen RC et al (2006), The genetics of plant metabolism. Nature Genetics 38, 842–849.

Alonso-Blanco C., Peeters, A. J. and Koornneef, M. (2006) Development of an AFLP based linkage map of Ler, Col and Cvi Arabidopsis thaliana ecotypes and construction of a Ler/Cvi recombinant inbred line population. Plant J. 14(2), 259–271.

See Also

multitrait

Examples

## Not run: 
    data(multitrait)
    data(locations)
    multiloc <- addloctocross(multitrait,locations)
    results <- scanall(multiloc)
    mqmplot.cistrans(results,multiloc, 5, FALSE, TRUE)

## End(Not run)

[Package qtl version 1.74 Index]