makeqtl {qtl}	R Documentation

Make a qtl object

Description

This function takes a cross object and specified chromosome numbers and positions and pulls out the genotype probabilities or imputed genotypes at the nearest pseudomarkers, for later use by the function fitqtl.

Usage

makeqtl(cross, chr, pos, qtl.name, what=c("draws","prob"))

Arguments

cross
An object of class cross. See read.cross for details.

chr
Vector indicating the chromosome for each QTL. (These should be character strings referring to the chromosomes by name.)

pos
Vector (of same length as chr) indicating the positions on the chromosome to be taken. If there is no marker or pseudomarker at a position, the nearest position is used.

qtl.name
Optional user-specified name for each QTL, used in the drop-one-term ANOVA table in fitqtl. If unspecified, the names will be of the form "Chr1@10" for a QTL on Chromsome 1 at 10 cM.

what
Indicates whether to pull out the imputed genotypes or the genotype probabilities.

Details

This function will take out the genotype probabilities and imputed genotypes if they are present in the input cross object. If both fields are missing in the input object, the function will report an error. Before running this function, the user must have first run either sim.geno (for what="draws") or calc.genoprob (for what="prob").

Value

An object of class qtl with the following elements (though only one of geno and prob will be included, according to whether what is given as "draws" or "prob"):

geno
Imputed genotypes.

prob
Genotype probabilities.

name
User-defined name for each QTL, or a name of the form "Chr1@10".

altname
QTL names of the form "Q1", "Q2", etc.

chr
Input vector of chromosome numbers.

pos
Input vector of chromosome positions.

n.qtl
Number of QTLs.

n.ind
Number of individuals.

n.gen
A vector indicating the number of genotypes for each QTL.

Author(s)

Hao Wu; Karl W Broman, broman@wisc.edu

See Also

fitqtl, calc.genoprob, sim.geno, dropfromqtl, replaceqtl, addtoqtl, summary.qtl, reorderqtl

Examples

data(fake.f2)

# take out several QTLs and make QTL object
qc <- c("1", "6", "13")
qp <- c(25.8, 33.6, 18.63)
fake.f2 <- subset(fake.f2, chr=qc)

fake.f2 <- sim.geno(fake.f2, n.draws=8, step=2, err=0.001)
qtl <- makeqtl(fake.f2, qc, qp, what="draws")
summary(qtl)

[Package qtl version 1.74 Index]