map10 {qtl}	R Documentation

An example genetic map

Description

A genetic map corresponding approximately to the mouse genome with a 10 cM marker spacing.

Usage

data(map10)

Format

An object of class map: a list whose components are vectors of marker locations. This map approximates the mouse genome, with 20 chromosomes (including the X chromosome) and 187 markers at an approximately 10 cM spacing. The markers are equally spaced on each chromosome, but the spacings are a bit above or below 10 cM, so that the lengths match those in the Mouse Genome Database.

See Also

sim.map, plotMap, pull.map

Examples

  data(map10)
  plot(map10)

  mycross <- sim.cross(map10, type="f2", n.ind=100)

[Package qtl version 1.74 Index]