map2table {qtl}	R Documentation

Convert genetic map from list to table.

Description

Convert a map object (as a list) to a table (as a data frame).

Usage

map2table(map, chr)

Arguments

map
A map object: a list whose components (corresponding to chromosomes) are either vectors of marker positions or matrices with two rows of sex-specific marker positions.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

Value

A data frame with two or three columns: chromosome and sex-averaged position, or chromosome, female position, and male position.

The row names are the marker names.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

table2map, pull.map, est.map

Examples

data(fake.f2)
map <- pull.map(fake.f2)
map_as_tab <- map2table(map)

[Package qtl version 1.74 Index]