mapthis {qtl}	R Documentation

Simulated data for illustrating genetic map construction

Description

Simulated data for an F2 intercross, obtained using sim.cross, useful for illustrating the process of constructing a genetic map.

Usage

data(mapthis)

Format

An object of class cross. See read.cross for details.

Details

These are simulated data, consisting of 300 F2 individuals typed at 100 markers on five chromosomes. There are no real phenotypes, just a set of individual identifiers. The data were simulated for the purpose of illustrating the process of constructing a genetic map. The markers are all assigned to a single chromosome and in a random order, and there are a number of problematic markers and individuals.

See https://rqtl.org/tutorials/geneticmaps.pdf for a tutorial on how to construct a genetic map with these data.

Author(s)

Karl W Broman, broman@wisc.edu

References

Broman, K. W. (2010) Genetic map construction with R/qtl. Technical report #214, Department of Biostatistics and Medical Informatics, University of Wisconsin–Madison

See Also

fake.f2, est.rf, est.map, formLinkageGroups, orderMarkers

Examples

data(mapthis)
summary(mapthis)
plot(mapthis)

[Package qtl version 1.74 Index]