markerlrt {qtl}	R Documentation

General likelihood ratio test for association between marker pairs

Description

Calculate a LOD score for a general likelihood ratio test for each pair of markers, to assess their association.

Usage

markerlrt(cross)

Arguments

cross
An object of class cross. See read.cross for details.

Value

The input cross object is returned with a component, rf, added. This is a matrix of size (tot.mar x tot.mar). The diagonal contains the number of typed meioses per marker, the upper and lower triangles each contain the LOD scores.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

plotRF, est.rf, badorder

Examples

data(badorder)
badorder <- markerlrt(badorder)
plotRF(badorder)

[Package qtl version 1.74 Index]