markernames {qtl}	R Documentation

Pull out the marker names from a cross

Description

Pull out the marker names from a cross object as one big vector.

Usage

markernames(cross, chr)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

Value

A vector of character strings (the marker names).

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.map, phenames, chrnames

Examples

data(listeria)
markernames(listeria, chr=5)

[Package qtl version 1.74 Index]