max.scanPhyloQTL {qtl}	R Documentation

Maximum peak in genome scan to map a QTL to a phylogenetic tree

Description

Print the chromosome with the maximum LOD score across partitions, from the results of scanPhyloQTL.

Usage

## S3 method for class 'scanPhyloQTL'
max(object, chr, format=c("postprob", "lod"),
        ...)

Arguments

object
An object output by the function scanPhyloQTL.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

format
Indicates whether to provide LOD scores or approximate posterior probabilities; see the help file for summary.scanPhyloQTL.

...
Ignored at this point.

Details

The output, and the use of the argument format, is as in summary.scanPhyloQTL.

Value

An object of class summary.scanPhyloQTL, to be printed by print.summary.scanPhyloQTL.

Author(s)

Karl W Broman, broman@wisc.edu

References

Broman, K. W., Kim, S., An\'e, C. and Payseur, B. A. Mapping quantitative trait loci to a phylogenetic tree. In preparation.

See Also

scanPhyloQTL, plot.scanPhyloQTL, summary.scanPhyloQTL, max.scanone, inferredpartitions, simPhyloQTL

Examples

## Not run: 
# example map; drop X chromosome
data(map10)
map10 <- map10[1:19]

# simulate data
x <- simPhyloQTL(4, partition="AB|CD", crosses=c("AB", "AC", "AD"),
                 map=map10, n.ind=150,
                 model=c(1, 50, 0.5, 0))

# run calc.genoprob on each cross
x <- lapply(x, calc.genoprob, step=2)

# scan genome, at each position trying all possible partitions
out <- scanPhyloQTL(x, method="hk")

# maximum peak
max(out, format="lod")

# approximate posterior probabilities at peak
max(out, format="postprob")

# all peaks above a threshold for LOD(best) - LOD(2nd best)
summary(out, threshold=1, format="lod")

# all peaks above a threshold for LOD(best), showing approx post'r prob
summary(out, format="postprob", threshold=3)

# plot of results
plot(out)

## End(Not run)

[Package qtl version 1.74 Index]