max.scanone {qtl}	R Documentation

Maximum peak in genome scan

Description

Print the row of the output from scanone that corresponds to the maximum LOD, genome-wide.

Usage

## S3 method for class 'scanone'
max(object, chr, lodcolumn=1, na.rm=TRUE, ...)

Arguments

object
An object of the form output by the function scanone: a data.frame whose third column is the LOD score.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

lodcolumn
An integer, indicating which of the LOD score columns should be considered in pulling out the peak (these are indexed 1, 2, ...).

na.rm
A logical indicating whether missing values should be removed.

...
Ignored.

Value

An object of class summary.scanone, to be printed by print.summary.scanone. This is a data.frame with one row, corresponding to the maximum LOD peak either genome-wide or for the particular chromosome specified.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scanone, plot.scanone, summary.scanone

Examples

data(listeria)

listeria <- calc.genoprob(listeria, step=2.5)
out <- scanone(listeria, model="2part", upper=TRUE)
# Maximum peak for LOD(p,mu)
max(out)

# Maximum peak for LOD(p,mu) on chr 5
max(out,chr=5)

# Maximum peak for LOD(p,mu) on chromosomes other than chr 13
max(out,chr="-13")

# Maximum peak for LOD(p)
max(out, lodcolumn=2)

# Maximum peak for LOD(mu)
max(out, lodcolumn=3)

[Package qtl version 1.74 Index]