max.scantwo {qtl}	R Documentation

Maximum peak in two-dimensional genome scan

Description

Print the pair of loci with the largest LOD score in the results of scantwo.

Usage

## S3 method for class 'scantwo'
max(object, lodcolumn=1,
    what=c("best", "full", "add", "int"),
    na.rm=TRUE, ...)

Arguments

object
An object of class scantwo, the output of the function scantwo.

lodcolumn
If the scantwo results contain LOD scores for multiple phenotypes, this argument indicates which to use.

what
Indicates for which LOD score the maximum should be reported.

na.rm
Ignored.

...
Ignored.

Details

This is very similar to the summary.scantwo function, though this pulls out one pair of positions.

If what="best", we find the pair of positions at which the LOD score for the full model (2 QTL + interaction) is maximized, and then also print the positions on that same pair of chromosomes at which the additive LOD score is maximized.

In the other cases, we pull out the pair of positions with the largest LOD score; which LOD score is considered is indicated by the what argument.

Value

An object of class summary.scantwo, to be printed by print.summary.scantwo, with the pair of positions with the maximum LOD score. (Which LOD score is considered is indicated by the what argument.)

Output of addpair

Note that, for output from addpair in which the new loci are indicated explicitly in the formula, the summary provided by max.scantwo is somewhat special.

All arguments (except, of course, the input object) are ignored.

If the formula is symmetric in the two new QTL, the output has just two LOD score columns: lod.2v0 comparing the full model to the model with neither of the new QTL, and lod.2v1 comparing the full model to the model with just one new QTL.

If the formula is not symmetric in the two new QTL, the output has three LOD score columns: lod.2v0 comparing the full model to the model with neither of the new QTL, lod.2v1b comparing the full model to the model in which the first of the new QTL is omitted, and lod.2v1a comparing the full model to the model with the second of the new QTL omitted.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scantwo, plot.scantwo, summary.scantwo

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2, step=10)
out.2dim <- scantwo(fake.f2, method="hk")
max(out.2dim)

[Package qtl version 1.74 Index]