movemarker {qtl}	R Documentation

Move a marker to a new chromosome

Description

Move a specified marker to a different chromosome.

Usage

movemarker(cross, marker, newchr, newpos)

Arguments

cross
An object of class cross. See read.cross for details.

marker
The name of the marker to be moved (a character string).

newchr
The chromosome to which the marker should be moved.

newpos
The position (in cM) at which the marker should be placed. If missing, the marker is placed at the end of the chromosome.

Value

The input cross object, but with the specified marker moved to the specified chromosome.

All intermediate calculations (such as the results of calc.genoprob and est.rf) are removed.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

switch.order

Examples

data(badorder)
badorder <- movemarker(badorder, "D2M937", 3, 48.15)
badorder <- movemarker(badorder, "D3M160", 2, 28.83)

[Package qtl version 1.74 Index]