mqmaugment {qtl}	R Documentation

MQM augmentation

Description

Fill in missing genotypes for MQM mapping. For each missing or incomplete marker it fills in (or ‘augments’) all possible genotypes, thus creating new candidate ‘individuals’. The probability of each indidual is calculated using information on neighbouring markers and recombination frequencies. When a genotype of an augmented genotype is less likely than the minprob parameter it is dropped from the dataset. The augmented list of individuals is returned in a new cross object. For a full discussion on augmentation see the MQM tutorial online.

Usage

mqmaugment(cross, maxaugind=82, minprob=0.1,
           strategy=c("default","impute","drop"),
           verbose=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

maxaugind
Maximum number of augmentations per individual. The default of 82 allows for six missing markers for an individual in a BC cross (2^6=64) and four missing markers in an F2 (3^4=81). When a large number of markers are missing this default number is quickly reached.

minprob
Return individuals with augmented genotypes that have at least this probability of occurring. minprob is a value between 0 and 1. For example a value of 0.5 will drop all genotypes that are half as likely as the most likely genotype (candidate of the individual). The default value of 0.1 will drop all genotypes that are less likely of ocurring than 1 in 10, compared against the most likely genotype. Use a value of 1.0 to return a single filled in genotype for each individual.

strategy
When individuals have too much missing data and augmentation fails three options are provided: 1. "default": Calculate genotypes at missing marker positions, accounting for minprob, and add this individual to the set. 2. "impute": Calculate the most likely genotypes at missing marker positions and impute maxaugind individual-variants around the most likely genotype. 3. "drop": Drop individuals that cannot be augmented from the dataset, this option is not advised because information from the dropped individuals will be lost.

verbose
If TRUE, give verbose output

Value

Returns the cross object with augmented individuals (many individuals from the data set will be repeated multiple times). Some individuals may have been dropped completely when the probability falls below minprob. An added component to the cross object named mqm contains information on exactly which individuals are retained and repeated.

Note

The sex chromosome 'X' is treated like autosomes during augmentation. With an F2 the sex chromosome is not considered. This will change in a future version of MQM. Run with verbose=TRUE to verify how many individuals are augmented versus moved to the second augmentation round. This could have an effect on the resulting dataset or check the return cross$mqm values. Compare results by using minprob=1.

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

fill.geno - Alternative routine for estimating missing data

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(map10)                    # Genetic map modeled after mouse

# simulate a cross (autosomes 1-10)
qtl <- c(3,15,1,0)             # QTL model: chr, pos'n, add've & dom effects
cross <- sim.cross(map10[1:10],qtl,n=100,missing.prob=0.01)

# MQM
crossaug <- mqmaugment(cross)  # Augmentation
cat(crossaug$mqm$Nind,'real individuals retained in dataset',
    crossaug$mqm$Naug,'individuals augmented\n')

result <- mqmscan(crossaug)    # Scan

# show LOD interval of the QTL on chr 3
lodint(result,chr=3)

[Package qtl version 1.74 Index]