mqmautocofactors {qtl}	R Documentation

Automatic setting of cofactors, taking marker density into account

Description

Sets cofactors, taking underlying marker density into account. Together with mqmscan cofactors are selected through backward elimination.

Usage

mqmautocofactors(cross, num=50, distance=5, dominance=FALSE, plot=FALSE, verbose=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

num
Number of cofactors to set (warns when setting too many cofactors).

distance
Minimal distance between two cofactors, in cM.

dominance
If TRUE, create a cofactor list that is safe to use with the dominance scan mode of MQM. See mqmscan for details.

plot
If TRUE, plots a genetic map displaying the selected markers as cofactors.

verbose
If TRUE, give verbose output.

Value

A list of cofactors to be used with mqmscan.

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

    data(hyper)                     # hyper dataset
    
    hyperfilled <- fill.geno(hyper)
    cofactors <- mqmautocofactors(hyperfilled,15)	# Set 15 Cofactors
    result <- mqmscan(hyperfilled,cofactors)	# Backward model selection
    mqmgetmodel(result)

[Package qtl version 1.74 Index]