mqmextractmarkers {qtl}	R Documentation

MQM marker extraction

Description

Extract the real markers from a cross object that includes pseudo markers

Usage

mqmextractmarkers(mqmresult)

Arguments

mqmresult
result from mqmscan, including pseudo markers

Value

Returns a scanone object with the pseudo markers removed

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

multitrait <- fill.geno(multitrait)

result <- mqmscan(multitrait)
newresult <- mqmextractmarkers(result)

[Package qtl version 1.74 Index]