mqmfind.marker {qtl}	R Documentation

Fetch significant markers after permutation analysis

Description

Fetch significant makers after permutation analysis. These markers can be used as cofactors for model selection in a forward stepwise approach.

Usage

mqmfind.marker(cross, mqmscan = NULL, perm = NULL, alpha = 0.05, verbose=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

mqmscan
Results from either scanone or mqmscan

perm
a scanoneperm object

alpha
Threshold value, everything with significance < alpha is reported

verbose
Display more output on verbose=TRUE

Value

returns a matrix with at each row a significant marker (determined from the scanoneperm object) and with columns: markername, chr and pos (cM)

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

mqmprocesspermutation - Function called to convert results from an mqmpermutation into an scanoneperm object

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

# Use the multitrait dataset
data(multitrait)

# Set cofactors at each 3th marker
cof <- mqmsetcofactors(multitrait,3)

# impute missing genotypes
multitrait <- fill.geno(multitrait)

# log transform the 7th phenotype
multitrait <- transformPheno(multitrait, 7)

# Bootstrap 50 runs in batches of 10
## Not run: result <- mqmpermutation(multitrait,scanfunction=mqmscan,cofactors=cof,
                         pheno.col=7,n.perm=50,batchsize=10)

## End(Not run)

# Create a permutation object
f2perm <- mqmprocesspermutation(result)

# What LOD score is considered significant ?
summary(f2perm)

# Find markers with a significant QTL effect (First run is original phenotype data)
marker <- mqmfind.marker(multitrait,result[[1]],f2perm)

# Print it to the screen
marker

[Package qtl version 1.74 Index]