mqmgetmodel {qtl}	R Documentation

Retrieve the QTL model used in mapping from the results of an MQM scan

Description

Retrieves the QTL model used for scanning from the output of an MQM scan. The model only contains the selected cofactors significant at the specified cofactor.significance from the results of an mqm scan

Usage

    mqmgetmodel(scanresult)

Arguments

scanresult
An object returned by mqmscan, including cofactors and QTL model.

Value

The function returns the multiple QTL model created, which consists of the cofactors selected during the modeling phase of the algorithm. This model was used when scanning for additional QTL in the mqmscan function. The format of the model is compatible with the makeqtl function. For more information about the format of the model see the makeqtl page. When no cofactor was selected in the modeling phase no model was created, then this function will return a NULL value.

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

mqmsetcofactors - Setting multiple cofactors for backward elimination

makeqtl - Make a qtl object

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

    data(hyper)
        
    hyperfilled <- fill.geno(hyper)
    cofactors <- mqmsetcofactors(hyperfilled,4)
    result <- mqmscan(hyperfilled,cofactors)
    mqmgetmodel(result)
    plot(mqmgetmodel(result))

[Package qtl version 1.74 Index]