mqmplot.cistrans {qtl}	R Documentation

cis-trans plot

Description

Plot results for a genomescan using a multiple-QTL model. With genetic location for the traits it is possible to show cis- and trans- locations, and detect trans-bands

Usage

mqmplot.cistrans(result, cross, threshold=5, onlyPEAK=TRUE,
                 highPEAK=FALSE, cisarea=10, pch=22, cex=0.5,
                 verbose=FALSE, ...)

Arguments

result
An object of class mqmmulti. See mqmscanall for details.

cross
An object of class cross. See read.cross for details.

threshold
Threshold value in LOD, Markers that have a LOD score above this threshold are plotted as small squares (see pch parameter). The markers with LODscores below this threshold are not visible

onlyPEAK
Plot only the peak markers ? (TRUE/FALSE) (Peak markers are markers that have a QTL likelihood above threshold and higher than other markers in the same region)

highPEAK
Highlight peak markers ? (TRUE/FALSE). When using this option peak markers (the marker with the highest LOD score in a region above the threshold gets an 25% increase in size and is displayed in red)

cisarea
Adjust the two green lines around the line y=x

pch
What kind of character is used in plotting of the figure (Default: 22, small square)

cex
Size of the points plotted (default to 0.5 half of the original size)

verbose
If TRUE, give verbose output

...
Extra parameters will be passed to points

Value

Plotting routine, so no return

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

data(locations)
multiloc <- addloctocross(multitrait,locations)
multiloc <- calc.genoprob(multiloc)
results <- scanall(multiloc, method="hk")
mqmplot.cistrans(results, multiloc, 5, FALSE, TRUE)

[Package qtl version 1.74 Index]