mqmplot.clusteredheatmap {qtl}	R Documentation

Plot clustered heatmap of MQM scan on multiple phenotypes

Description

Plot the results from a MQM scan on multiple phenotypes.

Usage

mqmplot.clusteredheatmap(cross, mqmresult, directed=TRUE, legend=FALSE,
                         Colv=NA, scale="none", verbose=FALSE,
                         breaks = c(-100,-10,-3,0,3,10,100),
                         col = c("darkblue","blue","lightblue","yellow",
                                 "orange","red"), ...)

Arguments

cross
An object of class cross. See read.cross for details.

mqmresult
Result object from mqmscanall, the object needs to be of class mqmmulti

directed
Take direction of QTLs into account (takes more time because of QTL direction calculations

legend
If TRUE, add a legend to the plot

Colv
Cluster only the Rows, the columns (Markers) should not be clustered

scale
character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default "none"

verbose
If TRUE, give verbose output.

breaks
Color break points for the LOD scores

col
Colors used between breaks

...
Additional arguments passed to heatmap.

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

multitrait <- fill.geno(multitrait) # impute missing genotype data
result <- mqmscanall(multitrait, logtransform=TRUE)
cresults <- mqmplot.clusteredheatmap(multitrait,result)
groupclusteredheatmap(multitrait,cresults,10)

[Package qtl version 1.74 Index]