mqmplot.cofactors {qtl}	R Documentation

Plot cofactors on the genetic map

Description

Plots cofactors as created by mqmsetcofactors or mqmautocofactors on the genetic map.

Usage

mqmplot.cofactors(cross,cofactors, ...)

Arguments

cross
An object of class cross. See read.cross for details.

cofactors
List of cofactors to be analysed in the QTL model. To set cofactors use mqmautocofactors or mqmsetcofactors

.

...
Passed to plot.qtl

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)
cof1 <- mqmsetcofactors(multitrait,20)
cof2 <- mqmsetcofactors(multitrait,10)
op <- par(mfrow=c(2,1))
mqmplot.cofactors(multitrait,cof1,col="blue")
mqmplot.cofactors(multitrait,cof2,col="blue")
op <- par(mfrow=c(1,1))

[Package qtl version 1.74 Index]