mqmplot.directedqtl {qtl}	R Documentation

Plot LOD*Effect curves of a multiple-QTL model

Description

Plot the LOD*Effect curve for a genome scan with a multiple-QTL model (the output of mqmscan).

Usage

mqmplot.directedqtl(cross, mqmresult, pheno.col=1, draw = TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

mqmresult
Results from mqmscan of type scanone

pheno.col
From which phenotype in the crossobject are the result calculated

draw
If TRUE, draw the figure.

Value

Returns a scanone object, with added the effectsign calculated internally by the function effect.scan. For more info on the scanone object see: scanone

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

#Simulated F2 Population
f2qtl <- c(3,15,1,0)                                    # QTL at chromosome 3
data(map10)                                             # Mouse genetic map

f2cross <- sim.cross(map10,f2qtl,n=100,type="f2")       # Simulate a F2 Cross
f2cross <- fill.geno(f2cross)                           # Fill in missing genotypes
f2result <- mqmscan(f2cross)                        # Do a MQM scan of the genome
mqmplot.directedqtl(f2cross,f2result)

[Package qtl version 1.74 Index]