mqmplot.heatmap {qtl}	R Documentation

Heatmap of a genome of MQM scan on multiple phenotypes

Description

Plotting routine to display a heatmap of results obtained from a multiple-QTL model on multiple phenotypes (the output of mqmscanall)

Usage

mqmplot.heatmap(cross, result, directed=TRUE, legend=FALSE, breaks =
c(-100,-10,-3,0,3,10,100), col =
c("darkblue","blue","lightblue","yellow","orange","red"), ...)

Arguments

cross
An object of class cross. See read.cross for details.

result
Result object from mqmscanall, the object needs to be of class mqmmulti

directed
Take direction of QTLs into account (takes more time because of QTL direction calculations

legend
If TRUE, add a legend to the plot

breaks
Color break points for the LOD scores

col
Colors used between breaks

...
Additional arguments passed to the image function

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

multitrait <- fill.geno(multitrait) # impute missing genotype data
result <- mqmscanall(multitrait, logtransform=TRUE)
mqmplot.heatmap(multitrait,result)

[Package qtl version 1.74 Index]