mqmplot.multitrait {qtl}	R Documentation

Plot the results from a genomescan using a multiple-QTL model on multiple phenotypes

Description

Plotting routine to display the results from a multiple-QTL model on multiple phenotypes. It supports four different visualizations: a contourmap, heatmap, 3D graph or a multiple QTL plot created by using plot.scanone on the mqmmulti object

Usage

mqmplot.multitrait(result, type=c("lines","image","contour","3Dplot"),
                   group=NULL, meanprofile=c("none","mean","median"),
                   theta=30, phi=15, ...)

Arguments

result
Result object from mqmscanall

type
Selection of the plot method to visualize the data: "lines" (defaut plotting option), "image", "contour" and "3Dplot"

group
A numeric vector indicating which traits to plot. NULL means no grouping

meanprofile
Plot a mean/median profile from the group selected

theta
Horizontal axis rotation in a 3D plot

phi
Vertical axis rotation in a 3D plot

...
Additional arguments passed to plot.

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

multitrait <- fill.geno(multitrait) # impute missing genotype data
result <- mqmscanall(multitrait, logtransform=TRUE)
mqmplot.multitrait(result,"lines")
mqmplot.multitrait(result,"contour")
mqmplot.multitrait(result,"image")
mqmplot.multitrait(result,"3Dplot")

[Package qtl version 1.74 Index]