mqmprocesspermutation {qtl}	R Documentation

Convert mqmmulti objects into a scanoneperm object

Description

Function to convert mqmmulti objects into a scanoneperm object, this allows the use of R/qtl methods for permutation analysis that do not support the output of a multiple QTL scan using mqm's outputstructure.

Usage

mqmprocesspermutation(mqmpermutationresult = NULL)

Arguments

mqmpermutationresult
mqmmulti object obtained after performing permutations on a single trait.using the function mqmpermutation

Value

Output of the algorithm is a scanoneperm object. See also: summary.scanoneperm

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

# QTL at chromosome 3
f2qtl <- c(3,15,1,0)

# Mouse genetic map
data(map10)

# Simulate a F2 Cross
f2cross <- sim.cross(map10,f2qtl,n=100,type="f2")
## Not run: # Bootstrap MQM mapping on the f2cross
f2result <- mqmpermutation(f2cross,scanfunction=mqmscan)

## End(Not run)

# Create a permutation object
f2perm <- mqmprocesspermutation(f2result)

# What LOD score is considered significant?
summary(f2perm)

[Package qtl version 1.74 Index]