mqmscanfdr {qtl}	R Documentation

Estimate FDR for multiple trait QTL analysis

Description

Estimate the false discovery rate (FDR) for multiple trait analysis

Usage

mqmscanfdr(cross, scanfunction=mqmscanall,
  thresholds=c(1,2,3,4,5,7,10,15,20), n.perm=10,
  verbose=FALSE, ...
  )

Arguments

cross
An object of class cross. See read.cross for details.

scanfunction
QTL mapping function, Note: Must use scanall or mqmscanall. Otherwise this will not produce usefull results. Reason: We need a function that maps all traits ecause of the correlation structure which is not changed (between traits) during permutation (Valis options: scanall or mqmscanall)

thresholds
False discovery rate (FDR) is calculated for peaks above these LOD thresholds (DEFAULT=Range from 1 to 20, using 10 thresholds) Parameter is a list of LOD scores at which FDR is calculated.

n.perm
Number of permutations (DEFAULT=10 for quick analysis, however for publications use 1000, or higher)

verbose
verbose output

...
Parameters passed to the mapping function

Details

This function wraps the analysis of scanone, cim and mqmscan to scan for QTL in shuffled/randomized data. It is recommended to also install the snow library for parallelization of calculations. The snow library allows calculations to run on multiple cores or even scale it up to an entire cluster, thus speeding up calculation by the number of computers used.

Value

Returns a data.frame with 3 columns: FalsePositives, FalseNegatives and False Discovery Rates. In the rows the userspecified thresholds are with scores for the 3 columns.

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

References

Bruno M. Tesson, Ritsert C. Jansen (2009) Chapter 3.7. Determining the significance threshold eQTL Analysis in Mice and Rats 1, 20–25

Churchill, G. A. and Doerge, R. W. (1994) Empirical threshold values for quantitative trait mapping. Genetics 138, 963–971.

Rossini, A., Tierney, L., and Li, N. (2003), Simple parallel statistical computing. R. UW Biostatistics working paper series University of Washington. 193

Tierney, L., Rossini, A., Li, N., and Sevcikova, H. (2004), The snow Package: Simple Network of Workstations. Version 0.2-1.

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

# impute missing genotype data
multitrait <- fill.geno(multitrait)
## Not run: # Calculate the thresholds
result <- mqmscanfdr(multitrait, threshold=10.0, n.perm=1000)

## End(Not run)

[Package qtl version 1.74 Index]