mqmsetcofactors {qtl}	R Documentation

Set cofactors at fixed intervals, to be used with MQM

Description

Set cofactors, at fixed marker intervals. Together with mqmscan cofactors are selected through backward elimination.

Usage

mqmsetcofactors(cross, each = NULL, cofactors=NULL, sexfactors=NULL, verbose=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

each
Every 'each' marker will be used as a cofactor, when each is used the cofactors and sexfactors parameter is ignored

cofactors
List of cofactors to be analysed in the QTL model. To set cofactors use mqmautocofactors or mqmsetcofactors; when each is set, this parameter is ignored

sexfactors
list of markers which should be treated as dominant cofactors (sexfactors), when each is set, this parameter is ignored

verbose
If TRUE, print tracing information.

Value

An list of cofactors to be passed into mqmscan.

Author(s)

Ritsert C Jansen; Danny Arends; Pjotr Prins; Karl W Broman broman@wisc.edu

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

    data(hyper)                                 # Hyper dataset
    
    hyperfilled <- fill.geno(hyper)
  # Automatic cofactors every third marker
    cofactors <- mqmsetcofactors(hyperfilled,3)
    result <- mqmscan(hyperfilled,cofactors)	# Backward model selection
    mqmgetmodel(result)
  #Manual cofactors at markers 3,6,9,12,40 and 60
  cofactors <- mqmsetcofactors(hyperfilled,cofactors=c(3,6,9,12,40,60))
    result <- mqmscan(hyperfilled,cofactors)	# Backward model selection
    mqmgetmodel(result)

[Package qtl version 1.74 Index]