mqmtestnormal {qtl}	R Documentation

Shapiro normality test used for MQM

Description

Wraps a shapiro's normality test from the nortest package. This function is used in MQM to test the normality of the trait under investigation

Usage

  mqmtestnormal(cross, pheno.col = 1,significance=0.05, verbose=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

pheno.col
Column number in the phenotype matrix which should be used as the phenotype. This can be a vector of integers.

significance
Significance level used in the normality test. Lower significance levels will accept larger deviations from normality.

verbose
If TRUE, print result as well as return it.

Details

For augmented data (as from mqmaugment), the cross is first reduced to distinct individuals. Furthermore the shapiro used to test normality works only for 3 <= nind(cross) <= 5000

Value

Boolean indicating normality of the trait in pheno.col. (FALSE when not normally distributed.)

Author(s)

Danny Arends danny.arends@gmail.com

See Also

shapiro.test - Function wrapped by our mqmtestnormal

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)

# test normality of 7th phenotype
mqmtestnormal(multitrait, pheno.col=7)

# take log
multitrait <- transformPheno(multitrait, pheno.col=7, transf=log)

# test again
mqmtestnormal(multitrait, pheno.col=7)

[Package qtl version 1.74 Index]