nmissing {qtl}	R Documentation

Number of missing genotypes

Description

Count the number of missing genotypes for each individual or each marker in a cross.

Usage

nmissing(cross, what=c("ind","mar"))

Arguments

cross
An object of class cross. See read.cross for details.

what
Indicates whether to count missing genotypes for each individual or each marker.

Value

A vector containing the number of missing genotypes for each individual or for each marker.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

ntyped, summary.cross, nind, totmar

Examples

data(listeria)

# plot number of missing genotypes for each individual
plot(nmissing(listeria))

# plot number of missing genotypes for each marker
plot(nmissing(listeria, what="mar"))

[Package qtl version 1.74 Index]