nqrank {qtl}	R Documentation

Transform a vector of quantitative values to the corresponding normal quantiles

Description

Transform a vector of quantitative values to the corresponding normal quantiles (preserving the mean and SD).

Usage

nqrank(x, jitter)

Arguments

x
A numeric vector

jitter
If TRUE, randomly jitter the values to break ties.

Value

A numeric vector; the input x is converted to ranks and then to normal quantiles.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

rank, qnorm, transformPheno

Examples

data(hyper)

hyper <- transformPheno(hyper, pheno.col=1, transf=nqrank)

[Package qtl version 1.74 Index]