plot.comparegeno {qtl}	R Documentation

Plot genotype comparison

Description

Plot the results of the comparison of all pairs of individuals' genotypes. A histogram of the proportion of matching genotypes, with tick marks at individual values below, via rug.

Usage

## S3 method for class 'comparegeno'
plot(x, breaks=NULL, main="",
     xlab="Proportion matching genotypes", ...)

Arguments

x
An object of class "comparegeno", as produced by comparegeno.

breaks
Passed to hist, with the default 2 √{n} where n is the number of pairs of individuals.

main
Title for the plot.

xlab
x-axis label for the plot.

...
Passed to hist.

Details

Creates a histogram with hist with ticks at individual values using rug.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

comparegeno, summary.comparegeno

Examples

data(fake.f2)
cg <- comparegeno(fake.f2)
plot(cg)

[Package qtl version 1.74 Index]