plot.cross {qtl}	R Documentation

Plot various features of a cross object

Description

Plots grid of the missing genotypes, genetic map, and histograms or barplots of phenotypes for the data from an experimental cross.

Usage

## S3 method for class 'cross'
plot(x, auto.layout=TRUE, pheno.col,
          alternate.chrid=TRUE, ...)

Arguments

x
An object of class cross. See read.cross for details.

auto.layout
If TRUE, par(mfrow) is set so that all plots fit within one figure.

pheno.col
Vector of numbers or character strings corresponding to phenotypes that should be plotted. If unspecified, all phenotypes are plotted.

alternate.chrid
If TRUE and more than one chromosome is plotted, alternate the placement of chromosome axis labels, so that they may be more easily distinguished.

...
Ignored at this point.

Details

Calls plotMissing, plotMap and plotPheno to plot the missing genotypes, genetic map, and histograms or barplots of all phenotypes.

If auto.format=TRUE, par(mfrow) is used with ceiling(sqrt(n.phe+2)) rows and the minimum number of columns so that all plots fit on the plotting device.

Numeric phenotypes are displayed as histograms or barplots by calling plotPheno.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu; Brian Yandell

See Also

plotMissing, plotMap, plotPheno

Examples

data(fake.bc)
plot(fake.bc)

[Package qtl version 1.74 Index]