plotErrorlod {qtl}	R Documentation

Plot grid of error LOD values

Description

Plot a grid of the LOD scores indicating which genotypes are likely to be in error.

Usage

plotErrorlod(x, chr, ind, breaks=c(-Inf,2,3,4.5,Inf),
              col=c("white","gray85","hotpink","purple3"),
              alternate.chrid=FALSE, ...)

Arguments

x
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to be drawn in the plot. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

ind
Indicates the individuals for which the error LOD scores should be plotted (passed to subset.cross).

breaks
A set of breakpoints for the colors; must give one more breakpoint than color. Intervals are open on the left and closed on the right, except for the lowest interval.

col
A vector of colors to appear in the image.

alternate.chrid
If TRUE and more than one chromosome is plotted, alternate the placement of chromosome axis labels, so that they may be more easily distinguished.

...
Ignored at this point.

Details

Uses image to plot a grid with different shades of pixels to indicate which genotypes are likely to be in error.

Darker pixels have higher error LOD scores: LOD ≤ 2 in white; 2 < LOD <= 3 in gray; 3 < LOD <= 4.5 in pink; LOD > 4.5 in purple.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

References

Lincoln, S. E. and Lander, E. S. (1992) Systematic detection of errors in genetic linkage data. Genomics 14, 604–610.

See Also

calc.errorlod, top.errorlod, image, subset.cross, plotGeno

Examples

data(hyper)

# Calculate error LOD scores
hyper <- calc.errorlod(hyper,error.prob=0.01)

# plot the error LOD scores; print those above a specified cutoff
plotErrorlod(hyper)
plotErrorlod(hyper,chr=1)

[Package qtl version 1.74 Index]