plotPheno {qtl}	R Documentation

Plot a phenotype distribution

Description

Plots a histogram or barplot of the data for a phenotype from an experimental cross.

Usage

plotPheno(x, pheno.col=1, ...)

Arguments

x
An object of class cross. See read.cross for details.

pheno.col
The phenotype column to plot: a numeric index, or the phenotype name as a character string. Alternatively, one may give a numeric vector of phenotypes, in which case it must have the length equal to the number of individuals in the cross, and there must be either non-integers or values < 1 or > no. phenotypes; this last case may be useful for studying transformations.

...
Passed to hist or barplot.

Details

Numeric phenotypes are displayed as histograms with approximately 2 sqrt(n) bins. Phenotypes that are factors or that have very few unique values are displayed as barplots.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

plot.cross, plotMap, plotMissing, hist, barplot

Examples

data(fake.bc)
plotPheno(fake.bc, pheno.col=1)
plotPheno(fake.bc, pheno.col=3)
plotPheno(fake.bc, pheno.col="age")

[Package qtl version 1.74 Index]