plot.qtl {qtl}	R Documentation

Plot QTL locations

Description

Plot the locations of the QTL against a genetic map

Usage

## S3 method for class 'qtl'
plot(x, chr, horizontal=FALSE, shift=TRUE,
     show.marker.names=FALSE, alternate.chrid=FALSE, justdots=FALSE,
     col="red", ...)

Arguments

x
An object of class "qtl", as produced by makeqtl.

chr
Optional vector indicating the chromosomes to plot. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

horizontal
Specifies whether the chromosomes should be plotted horizontally.

shift
If TRUE, shift the first marker on each chromosome to be at 0 cM.

show.marker.names
If TRUE, marker names are included.

alternate.chrid
If TRUE and more than one chromosome is plotted, alternate the placement of chromosome axis labels, so that they may be more easily distinguished.

justdots
If FALSE, just plot dots at the QTL, rather than arrows and QTL names.

col
Color used to plot indications of QTL

...
Passed to plotMap.

Details

Creates a plot, via plotMap, and indicates the locations of the QTL in the input QTL object, x.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

plotMap, makeqtl

Examples

data(fake.f2)

# take out several QTLs and make QTL object
qc <- c("1", "6", "13")
qp <- c(25.8, 33.6, 18.63)
fake.f2 <- subset(fake.f2, chr=qc)

fake.f2 <- calc.genoprob(fake.f2, step=2, err=0.001)
qtl <- makeqtl(fake.f2, qc, qp, what="prob")
plot(qtl)
plot(qtl, justdots=TRUE, col="seagreen")

[Package qtl version 1.74 Index]