plot.rfmatrix {qtl}	R Documentation

Plot recombination fractions or LOD scores for a single marker

Description

Plot a slice (corresponding to a single marker) through the pairwise recombination fractions or LOD scores calculated by est.rf and extracted with pull.rf.

Usage

## S3 method for class 'rfmatrix'
plot(x, marker, ...)

Arguments

x
An object of class rfmatrix, as output by pull.rf.

marker
A single marker name, as a character string.

...
Optional arguments passed to plot.scanone.

Value

An object of class "scanone" (as output by scanone, and which may be summarized by summary.scanone or plotted with plot.scanone), containing the estimated recombination fractions or LOD scores for the input marker against all others.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.rf, est.rf, plotRF

Examples

data(fake.f2)

fake.f2 <- est.rf(fake.f2)
marker <- markernames(fake.f2, chr=5)[6]
lod <- pull.rf(fake.f2, "lod")
plot(lod, marker, bandcol="gray70")

[Package qtl version 1.74 Index]