plot.scanPhyloQTL {qtl}	R Documentation

Plot LOD curves from single-QTL scan to map QTL to a phylogenetic tree

Description

Plot the LOD curves for each partition for a genome scan with a single diallelic QTL (the output of scanPhyloQTL).

Usage

## S3 method for class 'scanPhyloQTL'
plot(x, chr, incl.markers=TRUE,
     col, xlim, ylim, lwd=2, gap=25, mtick=c("line", "triangle"),
     show.marker.names=FALSE, alternate.chrid=FALSE, legend=TRUE, ...)

Arguments

x
An object of class "scanPhyloQTL", as output by scanPhyloQTL.

chr
Optional vector indicating the chromosomes to plot. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

incl.markers
Indicate whether to plot line segments at the marker locations.

col
Optional vector of colors to use for each partition.

xlim
Limits for x-axis (optional).

ylim
Limits for y-axis (optional).

lwd
Line width.

gap
Gap separating chromosomes (in cM).

mtick
Tick mark type for markers (line segments or upward-pointing triangels).

show.marker.names
If TRUE, show the marker names along the x axis.

alternate.chrid
If TRUE and more than one chromosome is plotted, alternate the placement of chromosome axis labels, so that they may be more easily distinguished.

legend
Indicates whether to include a legend in the plot.

...
Passed to the function plot.scanone when it is called.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

References

Broman, K. W., Kim, S., An\'e, C. and Payseur, B. A. Mapping quantitative trait loci to a phylogenetic tree. In preparation.

See Also

scanPhyloQTL, max.scanPhyloQTL, summary.scanPhyloQTL, plot.scanone, inferredpartitions, simPhyloQTL, par, colors

Examples

## Not run: 
# example map; drop X chromosome
data(map10)
map10 <- map10[1:19]

# simulate data
x <- simPhyloQTL(4, partition="AB|CD", crosses=c("AB", "AC", "AD"),
                 map=map10, n.ind=150,
                 model=c(1, 50, 0.5, 0))

# run calc.genoprob on each cross
x <- lapply(x, calc.genoprob, step=2)

# scan genome, at each position trying all possible partitions
out <- scanPhyloQTL(x, method="hk")

# maximum peak
max(out, format="lod")

# approximate posterior probabilities at peak
max(out, format="postprob")

# all peaks above a threshold for LOD(best) - LOD(2nd best)
summary(out, threshold=1, format="lod")

# all peaks above a threshold for LOD(best), showing approx post'r prob
summary(out, format="postprob", threshold=3)

# plot of results
plot(out)

## End(Not run)

[Package qtl version 1.74 Index]