plot.scanoneboot {qtl}	R Documentation

Plot results of bootstrap for QTL position

Description

Plot a histogram of the results of a nonparametric bootstrap to assess uncertainty in QTL position.

Usage

## S3 method for class 'scanoneboot'
plot(x, ...)

Arguments

x
An object of class "scanoneboot", as output by scanoneboot.

...
Passed to the function hist when it is called.

Details

The function plots a histogram of the bootstrap results obtained by scanoneboot. Genetic marker locations are displayed by vertical lines at the bottom of the plot.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scanone, summary.scanoneboot

Examples

data(fake.f2)
fake.f2 <- calc.genoprob(fake.f2, step=1)

## Not run: out.boot <- scanoneboot(fake.f2, chr=13, method="hk")

summary(out.boot)
plot(out.boot)

[Package qtl version 1.74 Index]