plot.scanoneperm {qtl}	R Documentation

Plot permutation results for a single-QTL genome scan

Description

Plot a histogram of the permutation results from a single-QTL genome scan.

Usage

## S3 method for class 'scanoneperm'
plot(x, lodcolumn=1, ...)

Arguments

x
An object of class "scanoneperm", as output by scanone when n.perm is specified.

lodcolumn
This indicates the LOD score column to plot. This should be a single number between 1 and the number of LOD columns in the object input.

...
Passed to the function hist when it is called.

Details

The function plots a histogram of the permutation results obtained by scanone when n.perm is specified. If separate permutations were performed for the autosomes and the X chromosome (using perm.Xsp=TRUE), separate histograms are given.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scanone, summary.scanoneperm

Examples

data(fake.bc)
fake.bc <- calc.genoprob(fake.bc)

operm <- scanone(fake.bc, method="hk", n.perm=100)
plot(operm)

[Package qtl version 1.74 Index]