plot.scantwoperm {qtl}	R Documentation

Plot permutation results for a 2d, 2-QTL genome scan

Description

Plot a histogram of the permutation results from a two-dimensional, two-QTL genome scan.

Usage

## S3 method for class 'scantwoperm'
plot(x, lodcolumn=1, include_rug=TRUE, ...)

Arguments

x
An object of class "scantwoperm", as output by scantwo when n.perm is specified.

lodcolumn
This indicates the LOD score column to plot. This should be a single number between 1 and the number of LOD columns in the object input.

include_rug
If TRUE, include a call to rug.

...
Passed to the function hist when it is called.

Details

The function plots a histogram of the permutation results obtained by scantwo when n.perm is specified. Separate histograms are provided for the five LOD scores, full, fv1, int, add, and av1.

Value

None.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scantwo, summary.scantwoperm

Examples

data(fake.bc)
fake.bc <- calc.genoprob(fake.bc)

operm2 <- scantwo(fake.bc, method="hk", n.perm=10)
plot(operm2)

[Package qtl version 1.74 Index]