pull.argmaxgeno {qtl}	R Documentation

Pull out the results of the Viterbi algorithm from a cross

Description

Pull out the results of argmax.geno from a cross as a matrix.

Usage

pull.argmaxgeno(cross, chr, include.pos.info=FALSE, rotate=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

include.pos.info
If TRUE, include columns with marker name, chromosmoe ID, and cM position. (If include.pos.info=TRUE, we take rotate=TRUE.)

rotate
If TRUE, return matrix with individuals as columns and positions as rows. If FALSE, rows correspond to individuals.

Value

A matrix containing numeric indicators of the inferred genotypes. Multiple chromosomes are pasted together.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.geno, pull.genoprob, pull.draws, argmax.geno

Examples

data(listeria)
listeria <- argmax.geno(listeria, step=1, stepwidth="max")
amg <- pull.argmaxgeno(listeria, chr=c(5,13), include.pos.info=TRUE, rotate=TRUE)
amg[1:5,1:10]

[Package qtl version 1.74 Index]