pull.draws {qtl}	R Documentation

Pull out the genotype imputations from a cross

Description

Pull out the results of sim.geno from a cross as an array.

Usage

pull.draws(cross, chr)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

Value

An array containing numeric indicators of the imputed genotypes. Multiple chromosomes are pasted together. The dimensions are individuals by positions by imputations

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.geno, pull.genoprob, pull.argmaxgeno, sim.geno

Examples

data(listeria)
listeria <- sim.geno(listeria, step=5, stepwidth="max", n.draws=8)
dr <- pull.draws(listeria, chr=c(5,13))
dr[1:20,1:10,1]

[Package qtl version 1.74 Index]