pull.geno {qtl}	R Documentation

Pull out the genotype data from a cross

Description

Pull out the genotype data from a cross object, as a single big matrix.

Usage

pull.geno(cross, chr)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

Value

A matrix of size n.ind x tot.mar. The raw genotype data in the input cross object, with the chromosomes pasted together.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.pheno, pull.map pull.draws, pull.genoprob, pull.argmaxgeno

Examples

data(listeria)
dat <- pull.geno(listeria)

# image of the genotype data
image(1:ncol(dat),1:nrow(dat),t(dat),ylab="Individuals",xlab="Markers",
      col=c("red","yellow","blue","green","violet"))
abline(v=cumsum(c(0,nmar(listeria)))+0.5)
abline(h=nrow(dat)+0.5)

[Package qtl version 1.74 Index]