pull.genoprob {qtl}	R Documentation

Pull out the genotype probabilities from a cross

Description

Pull out the results of calc.genoprob from a cross as a matrix.

Usage

pull.genoprob(cross, chr, omit.first.prob=FALSE,
              include.pos.info=FALSE, rotate=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

omit.first.prob
If TRUE, omit the probabilities for the first genotype at each position (since they sum to 1).

include.pos.info
If TRUE, include columns with marker name, genotype, chromosome ID, and cM position. (If include.pos.info=TRUE, we take rotate=TRUE.)

rotate
If TRUE, return matrix with individuals as columns and positions/genotypes as rows. If FALSE, rows correspond to individuals.

Value

A matrix containing genotype probabilities. Multiple chromosomes and the multiple genotypes at each position are pasted together.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.geno, pull.argmaxgeno, pull.draws, calc.genoprob

Examples

data(listeria)
listeria <- calc.genoprob(listeria, step=1, stepwidth="max")
pr <- pull.genoprob(listeria, chr=c(5,13), omit.first.prob=TRUE, include.pos.info=TRUE, rotate=TRUE)
pr[1:5,1:10]

[Package qtl version 1.74 Index]