pull.map {qtl}	R Documentation

Pull out the genetic map from a cross

Description

Pull out the map portion of a cross object.

Usage

pull.map(cross, chr, as.table=FALSE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

as.table
If TRUE, return the genetic map as a table with chromosome assignments and marker names. If FALSE, return the map as a "map" object.

Value

The genetic map: a list with each component containing the marker positions (in cM) for a chromosome. Each component has class A or X according to whether it is an autosome or the X chromosome. The components are either vectors of marker positions or, for a sex-specific map, 2-row matrices containing the female and male marker locations. The map itself is given class map.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

replace.map, plotMap, map2table

Examples

data(fake.f2)
map <- pull.map(fake.f2)
plot(map)

[Package qtl version 1.74 Index]