pull.pheno {qtl}	R Documentation

Pull out phenotype data from a cross

Description

Pull out selected phenotype data from a cross object, as a data frame or vector.

Usage

pull.pheno(cross, pheno.col)

Arguments

cross
An object of class cross. See read.cross for details.

pheno.col
A vector specifying which phenotypes to keep or discard. This may be a logical vector, a numeric vector, or a vector of character strings (for the phenotype names). If missing, the entire set of phenotypes is output.

Value

A data.frame with columns specifying phenotypes and rows specifying individuals. If there is just one phenotype, a vector (rather than a data.frame) is returned.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.geno, pull.map

Examples

data(listeria)
pull.pheno(listeria, "sex")

[Package qtl version 1.74 Index]