pull.rf {qtl}	R Documentation

Pull out recombination fractions or LOD scores from a cross object

Description

Pull out either the pairwise recombination fractions or the LOD scores, as calculated by est.rf, from a cross object.

Usage

pull.rf(cross, what=c("rf", "lod"), chr)

Arguments

cross
An object of class cross. See read.cross for details.

what
Indicates whether to pull out a matrix of estimated recombination fractions or a matrix of LOD scores.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

Value

An object of class "rfmatrix", which is a matrix of either estimated recombination fractions between all marker pairs or of LOD scores (for the test of rf=1/2) for all marker pairs.

The genetic map is included as an attribute.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

est.rf, plot.rfmatrix, plotRF

Examples

data(fake.f2)

fake.f2 <- est.rf(fake.f2)
rf <- pull.rf(fake.f2)
lod <- pull.rf(fake.f2, "lod")
plot(rf[1,], lod[1,], xlab="rec frac", ylab="LOD score")
marker <- markernames(fake.f2, chr=5)[6]
par(mfrow=c(2,1))
plot(rf, marker, bandcol="gray70")
plot(lod, marker, bandcol="gray70")

[Package qtl version 1.74 Index]