qtlversion {qtl}	R Documentation

Installed version of R/qtl

Description

Print the version number of the currently installed version of R/qtl.

Usage

qtlversion()

Value

A character string with the version number of the currently installed version of R/qtl.

Author(s)

Karl W Broman, broman@wisc.edu

Examples

  qtlversion()

[Package qtl version 1.74 Index]