readMWril {qtl}	R Documentation

Read data for 4- or 8-way RIL

Description

Data for a set of 4- or 8-way recombinant inbred lines (RIL) is read from a pair of comma-delimited files and converted into an object of class cross. We require chromosome assignments for the genetic markers, and assume that markers are in their correct order.

Usage

readMWril(dir="", rilfile, founderfile,
          type=c("ri4self", "ri4sib", "ri8self", "ri8selfIRIP1", "ri8sib", "bgmagic16"),
          na.strings=c("-","NA"), rotate=FALSE, ...)

Arguments

dir
Directory in which the data files will be found. In Windows, use forward slashes ("/") or double backslashes ("\\") to specify directory trees.

rilfile
Comma-delimited file for the RIL, in the "csv" format described in the help file for read.cross.

founderfile
File with founder strains' genotypes, in the same orientation as the rilfile, but with just marker names and the founders' marker genotypes.

type
The type of RIL.

na.strings
A vector of strings which are to be interpreted as missing values. For the csv formats, these are interpreted globally for the entire file, so missing value codes in phenotypes must not be valid genotypes, and vice versa. For the gary format, these are used only for the phenotype data.

rotate
If TRUE, the rilfile and founderfile are rotated (really transposed), with rows corresponding to markers and columns corresponding to individuals.

...
Additional arguments, passed to the function read.table in the case of csv and csvr formats. In particular, one may use the argument sep to specify the field separator (the default is a comma) and dec to specify the character used for the decimal point (the default is a period).

Details

The rilfile should include a phenotype cross containing character strings of the form ABCDEFGH, indicating the cross used to generate each RIL. The genotypes should be coded as integers (e.g., 1 and 2).

The founder strains in the founderfile should be the strains A, B, C, ..., as indicated in the cross phenotype.

The default arrangement of the files is to have markers as columns and individuals/founders as rows. If rotate=TRUE, do the opposite: markers as rows and individuals/founders as columns.

Value

An object of class cross; see the help file for read.cross for details.

An additional component crosses is included; this is a matrix indicating the crosses used to generate the RIL.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

read.cross, sim.cross

Examples

## Not run: 
ril <- read.cross("../Data", "ril_data.csv", "founder_geno.csv", "ri4self",
                  rotate=TRUE)
## End(Not run)

[Package qtl version 1.74 Index]