reduce2grid {qtl}	R Documentation

Reduce to a grid of pseudomarkers.

Description

For high-density marker data, rather than run scanone at both the markers and at a set of pseudomarkers, we reduce to just a set of evenly-spaced pseudomarkers

Usage

reduce2grid(cross)

Arguments

cross
An object of class cross. See read.cross for details.

Details

Genotype probabilities (from calc.genoprob) and/or imputations (from sim.geno) are subset to a grid of pseudomarkers.

This is so that, in the case of high-density markers, we can do the genome scan calculations at a smaller set of points (on an evenly-spaced grid, but not at the markers) to save computation time.

You need to first have run calc.genoprob and/or sim.geno, and you must use stepwidth="fixed".

When plotting results with plot.scanone, use incl.markers=FALSE, as the output of scanone won't include information about the marker locations and so will plot tick marks only at the first marker on each chromosome.

Value

The input cross object with included genotype probabilities or imputations subset to an evenly-spaced grid.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

calc.genoprob, sim.geno, scanone, plot.scanone

Examples

data(hyper)
hyper <- calc.genoprob(hyper, step=2)
hypersub <- reduce2grid(hyper)

## Not run: out <- scanone(hypersub)
plot(out, incl.markers=FALSE)
## End(Not run)

[Package qtl version 1.74 Index]