reorderqtl {qtl}	R Documentation

Reorder the QTL in a qtl object

Description

This function changes the order of the QTL in a QTL object.

Usage

reorderqtl(qtl, neworder)

Arguments

qtl
A qtl object, as created by makeqtl.

neworder
A vector containing the positive integers up to the number of QTL in qtl, indicating the new order for the QTL. If missing, the QTL are ordered by chromosome and then by their position within a chromosome.

Details

Everything in the input qtl is reordered except the altname component, which contains names of the form Q1, Q2, etc.

Value

The input qtl object, with the loci reordered.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

makeqtl, fitqtl, dropfromqtl, addtoqtl, replaceqtl

Examples

data(fake.f2)

# take out several QTLs and make QTL object
qc <- c(1, 6, 13)
qp <- c(25.8, 33.6, 18.63)
fake.f2 <- subset(fake.f2, chr=qc)

fake.f2 <- calc.genoprob(fake.f2)
qtl <- makeqtl(fake.f2, qc, qp, what="prob")

qtl <- reorderqtl(qtl, c(2,3,1))
qtl

qtl <- reorderqtl(qtl)
qtl

[Package qtl version 1.74 Index]