replace.map {qtl}	R Documentation

Replace the genetic map of a cross

Description

Replace the map portion of a cross object.

Usage

replace.map(cross, map)
## S3 method for class 'cross'
replacemap(object, map)

Arguments

cross
An object of class cross. See read.cross for details.

object
Same as cross.

map
A list containing the new genetic map. This must be the same length and with the same marker names as that contained in cross.

Value

The input cross object with the genetic map replaced by the input map. Maps for results from calc.genoprob, sim.geno and argmax.geno are also replaced, using interpolation if necessary.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

pull.map, est.map

Examples

data(fake.f2)

newmap <- est.map(fake.f2)
plotMap(fake.f2, newmap)
fake.f2 <- replace.map(fake.f2, newmap)

[Package qtl version 1.74 Index]