replacemap.scantwo {qtl}	R Documentation

Replace the genetic map in QTL mapping results with an alternate map

Description

Replace the positions of LOD scores in output from scantwo with values based on an alternative map (such as a physical map), with pseudomarker locations determined by linear interpolation.

Usage

## S3 method for class 'scantwo'
replacemap(object, map)

Arguments

object
An object of class "scantwo", as output by the function scantwo.

map
A list containing the alternative genetic map. All chromosomes in object should have corresponding chromosomes in map, and markers must be in the same order in the two maps. There must be at least two markers on each chromosome in map that appear in object.

Details

The positions of pseudomarkers are determined by linear interpolation between markers. In the case of pseudomarkers beyond the ends of the terminal markers on chromosomes, we use the overall lengths of the chromosome in object and map to determine the new spacing.

Value

The input object with the positions of LOD scores revised to match those in the input map.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

replacemap.cross, est.map, replacemap.scanone

Examples

data(hyper)

origmap <- pull.map(hyper)
newmap <- est.map(hyper)
hyper <- replacemap(hyper, newmap)
hyper <- calc.genoprob(hyper, step=0)
out <- scantwo(hyper, method="hk")
out.rev <- replacemap(out, origmap)

[Package qtl version 1.74 Index]