rescalemap {qtl}	R Documentation

Rescale genetic maps

Description

Rescale a genetic map by multiplying all positions by a constant

Usage

rescalemap(object, scale=1e-6)

Arguments

object
An object of class cross (see read.cross for details) or map (see sim.map for details).

scale
Scale factor by which all positions will be multiplied.

Details

This function is included particularly for the case that map positions in a cross object were provided in basepairs and one wishes to quickly convert them to Mbp or some other approximation of cM distances. (In the mouse, 1 cM is approximation 2 Mbp, so one might use scale=5e-7 in this function.)

Value

If the input is a map object, a map object is returned; if the input is a cross object, a cross object is returned. In either case, the positions of markers are simply multiplied by scale.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

replace.map, est.map

Examples

data(hyper)
rescaled <- rescalemap(hyper, scale=2)
plotMap(hyper, rescaled)

[Package qtl version 1.74 Index]