scanMQM(MQMpackage)	R Documentation

scanMQM - Multiple qtl mapping for F2 crosses

Description

Main multiQTL scanning function for the R/qtl package (for use with F2 populations)

Usage

scanMQM(cross = NULL, cofactors = NULL, REMLorML = 0, alfa = 0.02, em.iter = 1000, windowsize = 25, step.size = 5, step.min = -20, step.max = 220)

Arguments

cross 	 Cross object from R/QTL

cofactors 	 List of cofactors (0:no cofactor, 1 cofactor, 2 sexfactor

REMLorML 	 Reduced maximum likelyhood (0) or Maximum likelyhood (1)

alfa 	 Global p-value (0<= alfa <= 1)

em.iter 	 Max itterations of the EM algorithm

windowsize 	 Window size when mappingQTL locations

step.size 	 Step.size used for the generation of output QTL.

step.min 	 Begin this many Cm before the chromosome

step.max 	 Maximum length in Cm

Details

Using ML or Reduced ML the algorithm employs a forward- or backward selection strategy to identify QTL hotspot.

Value

No return value, the cross object will be updated with the new information.

Support

If u find anything wrong with the algorithm please notify me at: Danny.Arends@gmail.com

Note

~~further notes~~ section{Warning}{ } section{Support}{ If u find anything wrong with the algorithm please notify me at: Danny.Arends@gmail.com }

Author(s)

Danny Arends

References

R.C. Janssen, K. Brohman, Danny Arends

Examples

        library(qtl)
        qtl <- c(3,15,3,7)                                                      # QTL at chromosome 3
        data(map10)                                                                     # Mouse genome
        cross <- sim.cross(map10,qtl,n=100)                     # Simulate a Cross
        result <- scanMQM(cross)                                        # Do a forward MQM model selection

[Package MQMpackage version 1.0 Index]