scanonevar {qtl}	R Documentation

Genome scan for QTL affecting mean and/or variance

Description

Genome scan with a single QTL model for loci that can affect the variance as well as the mean.

Usage

scanonevar(cross, pheno.col=1, mean_covar=NULL, var_covar=NULL,
           maxit=25, tol=1e-6, quiet=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

pheno.col
Column number in the phenotype matrix which should be used as the phenotype. This must be a single value (integer index or phenotype name) or a numeric vector of phenotype values, in which case it must have the length equal to the number of individuals in the cross, and there must be either non-integers or values < 1 or > no. phenotypes; this last case may be useful for studying transformations.

mean_covar
Numeric matrix with covariates affecting the mean.

var_covar
Numeric matrix with covariates affecting the variances.

maxit
Maximum number of iterations in the algorithm to fit the model at a given position.

tol
Tolerance for convergence.

quiet
If FALSE, print some information about the course of the calculations.

Value

A data frame (with class "scanone", in the form output by scanone), with four columns: chromosome, position, the -log P-value for the mean effect, and the -log P-value for the effect on the variance. The result is given class "scanone"

Author(s)

Lars Ronnegard and Karl Broman

References

Ronnegard, L. and Valdar W. (2011) Detecting major genetic loci controlling phenotypic variability in experimental crosses. Genetics 188:435-447

Ronnegard, L. and Valdar W. (2012) Recent developments in statistical methods for detecting genetic loci affecting phenotypic variability. BMC Genetics 13:63

See Also

scanone, summary.scanone, calc.genoprob, summary.scanoneperm

Examples

data(fake.bc)

fake.bc <- calc.genoprob(fake.bc, step=2.5)
out <- scanonevar(fake.bc)
color <- c("slateblue", "violetred")
plot(out, lod=1:2, col=color, bandcol="gray80")
legend("topright", lwd=2, c("mean", "variance"), col=color)

# use format="allpeaks" to get summary for each of mean and variance
#  also consider format="tabByCol" or format="tabByChr"
summary(out, format="allpeaks")

# with sex and age as covariates
covar <- fake.bc$pheno[,c("sex", "age")]
out.cov <- scanonevar(fake.bc, mean_covar=covar, var_covar=covar)

[Package qtl version 1.74 Index]