scanonevar.meanperm {qtl}	R Documentation

Permutation test for mean effect in scanonevar

Description

Executes permutations of the genotypes in the mean-effect part of scanonevar

Usage

scanonevar.meanperm(cross, pheno.col=1, mean_covar=NULL, var_covar=NULL,
           maxit=25, tol=1e-6, n.mean.perm = 2, seed = 27517, quiet=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

pheno.col
Column number in the phenotype matrix which should be used as the phenotype. This must be a single value (integer index or phenotype name) or a numeric vector of phenotype values, in which case it must have the length equal to the number of individuals in the cross, and there must be either non-integers or values < 1 or > no. phenotypes; this last case may be useful for studying transformations.

mean_covar
Numeric matrix with covariates affecting the mean.

var_covar
Numeric matrix with covariates affecting the variances.

maxit
Maximum number of iterations in the algorithm to fit the model at a given position.

tol
Tolerance for convergence.

n.mean.perm
Numeric vector of length one indicates the number of permutations to execute.

seed
Numeric vector of length one indicates the random seed to start the permutations.

quiet
If FALSE, print some information about the course of the calculations.

Value

A vector of length n.mean.perm of the maximum negative log10 p-value that resulted from each permutation.

[Package qtl version 1.74 Index]