shiftmap {qtl}	R Documentation

Shift starting points in genetic maps

Description

Shift starting points in a genetic map to a set of defined positions

Usage

shiftmap(object, offset=0)

Arguments

object
An object of class cross (see read.cross for details) or map (see sim.map for details).

offset
Defines the starting position for each chromosome. This should be a single value (to be used for all chromosomes) or a vector with length equal to the number of chromosomes, defining individual starting positions for each chromosome. For a sex-specific map (as in a 4-way cross), we use the same offset for both the male and female maps.

Value

If the input is a map object, a map object is returned; if the input is a cross object, a cross object is returned. In either case, the positions of markers are shifted so that the starting positions are as in offset.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

replace.map, est.map

Examples

data(hyper)
shiftedhyper <- shiftmap(hyper, offset=0)
par(mfrow=c(1,2))
plotMap(hyper, shift=FALSE, alternate.chrid=TRUE)
plotMap(shiftedhyper, shift=FALSE, alternate.chrid=TRUE)

[Package qtl version 1.74 Index]