simFounderSnps {qtl}	R Documentation

Simulate founder SNPs for a multiple-strain RIL

Description

Simulate genotype data for the founding strains for a panel of multiple-strain RIL.

Usage

simFounderSnps(map, n.str=c("4","8"), pat.freq)

Arguments

map
A list whose components are vectors containing the marker locations on each of the chromosomes.

n.str
Number of founding strains (4 or 8).

pat.freq
Frequency of SNP genotype patterns in the founder (a vector of length n.str/2 + 1): (monoallelic, SNP unique to one founder, SNP present in 2 founders, [and, for the case of 8 founders, SNP in 3/8 founders, SNP in 4/8 founders].)

Details

The SNPs are simulated to be in linkage equilibrium.

Value

A vector of the same length as there are chromosomes in map, with each component being a matrix of 0's and 1's, of dim n.str x n.mar.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

sim.map, sim.cross

Examples

data(map10)
x <- simFounderSnps(map10, "8", c(0, 0.5, 0.2, 0.2, 0.1))

[Package qtl version 1.74 Index]