simulatemissingdata {qtl}	R Documentation

Simulates missing genotype data

Description

Simulate missing genotype data by removing some genotype data from the cross object

Usage

  simulatemissingdata(cross, percentage = 5)

Arguments

cross
An object of class cross. See read.cross for details.

percentage
How much of the genotype data do we need to randomly drop?

Value

An object of class cross with percentage

Author(s)

Danny Arends danny.arends@gmail.com

See Also

The MQM tutorial: https://rqtl.org/tutorials/MQM-tour.pdf

MQM - MQM description and references

mqmscan - Main MQM single trait analysis

mqmscanall - Parallellized traits analysis

mqmaugment - Augmentation routine for estimating missing data

mqmautocofactors - Set cofactors using marker density

mqmsetcofactors - Set cofactors at fixed locations

mqmpermutation - Estimate significance levels

scanone - Single QTL scanning

Examples

data(multitrait)
multitrait <- fill.geno(multitrait)
multimissing5 <- simulatemissingdata(multitrait,perc=5)
perc <- (sum(nmissing(multimissing5))/sum(ntyped(multimissing5)))

[Package qtl version 1.74 Index]