subset.map {qtl}	R Documentation

Subsetting chromosomes for a genetic map

Description

Pull out a specified set of chromosomes from a map object.

Usage

## S3 method for class 'map'
subset(x, ...)
## S3 method for class 'map'
x[...]

Arguments

x
A list whose components are vectors of marker locations.

...
Vector of chromosome indices.

Value

The input map object, but with only the specified subset of chromosomes.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

subset.cross

Examples

data(map10)
map10 <- subset(map10, chr=1:5)

# you can also use brackets
map10 <- map10[2:3]

[Package qtl version 1.74 Index]