subset.scanone {qtl}	R Documentation

Subsetting the results of a genome scan

Description

Pull out a specified set of chromosomes and/or LOD columns from scanone output.

Usage

## S3 method for class 'scanone'
subset(x, chr, lodcolumn, ...)

Arguments

x
An object of class scanone, output from scanone.

chr
Optional vector specifying which chromosomes to keep. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

lodcolumn
A vector specifying which LOD columns to keep (or, if negative), omit. These should be between 1 and the number of LOD columns in the input x.

...
Ignored at this point.

Value

The input scanone object, but with only the specified subset of the data.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scanone, scanone

Examples

data(fake.bc)

fake.bc <- calc.genoprob(fake.bc, step=2.5)
out <- scanone(fake.bc, method="hk", pheno.col=1:2)

summary(subset(out, chr=18:19), format="allpeaks")

[Package qtl version 1.74 Index]