subset.scanoneperm {qtl}	R Documentation

Subsetting permutation test results

Description

Pull out results for a specified set LOD columns from permutation results from scanone.

Usage

## S3 method for class 'scanoneperm'
subset(x, repl, lodcolumn, ...)
## S3 method for class 'scanoneperm'
x[repl, lodcolumn]

Arguments

x
Permutation results from scanone, run with n.perm>0.

repl
A vector specifying which permutation replicates to keep or (if negative) omit.

lodcolumn
A vector specifying which LOD columns to keep or (if negative) omit. These should be between 1 and the number of LOD columns in the input x.

...
Ignored at this point.

Value

The input scanone permutation results, but with only the specified subset of the data.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scanoneperm, scanone, c.scanoneperm, cbind.scanoneperm, rbind.scanoneperm

Examples

data(fake.bc)

fake.bc <- calc.genoprob(fake.bc, step=5)
operm <- scanone(fake.bc, method="hk", pheno.col=1:2, n.perm=25)
operm2 <- subset(operm, lodcolumn=2)

# alternatively
operm2alt <- operm[,2]

[Package qtl version 1.74 Index]