subset.scantwoperm {qtl}	R Documentation

Subsetting two-dimensional permutation test results

Description

Pull out results for a specified set LOD columns from permutation results from scantwo.

Usage

## S3 method for class 'scantwoperm'
subset(x, repl, lodcolumn, ...)
## S3 method for class 'scantwoperm'
x[repl, lodcolumn]

Arguments

x
Permutation results from scantwo, run with n.perm>0.

repl
A vector specifying which permutation replicates to keep or (if negative) omit. Ignored in case of X-chr specific permutations

lodcolumn
A vector specifying which LOD columns to keep or (if negative) omit. These should be between 1 and the number of LOD columns in the input x.

...
Ignored at this point.

Value

The input scantwo permutation results, but with only the specified subset of the data.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

summary.scantwoperm, scantwo, c.scantwoperm, rbind.scantwoperm

Examples

data(fake.bc)

fake.bc <- calc.genoprob(fake.bc, step=0)
operm <- scantwo(fake.bc, method="hk", pheno.col=1:2, n.perm=5)
operm2 <- subset(operm, lodcolumn=2)

# alternatively
operm2alt <- operm[,2]

[Package qtl version 1.74 Index]