summary.comparegeno {qtl}	R Documentation

Print pairs of individuals with similar genotype data.

Description

Prints a summary the output from comparegeno that includes pairs of individuals whose proportion of matching genotypes is above a chosen threshold.

Usage

## S3 method for class 'comparegeno'
summary(object, thresh=0.9, ...)

Arguments

object
An object of class comparegeno, the output of the function comparegeno.

thresh
Threshold on the proportion of matching genotypes.

...
Ignored at this point.

Value

A data frame with each row being a pair of individuals and columns including the individual identifiers (via getid, or just as numeric indexes) along with the proportion of matching genotypes.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

comparegeno, plot.comparegeno

Examples

data(fake.f2)
cg <- comparegeno(fake.f2)
summary(cg, 0.7)

[Package qtl version 1.74 Index]