summary.fitqtl {qtl}	R Documentation

Summary of fit of qtl model

Description

Print summary information about the results of fitqtl.

Usage

## S3 method for class 'fitqtl'
summary(object, pvalues=TRUE, simple=FALSE, ...)

Arguments

object
Output from fitqtl.

pvalues
If FALSE, don't include p-values in the summary.

simple
If TRUE, don't include p-values or sums of squares in the summary.

...
Ignored at this point.

Value

An object of class summary.fitqtl, which is not all that different than the input, but when printed gives summary information about the results.

Author(s)

Hao Wu; Karl W Broman, broman@wisc.edu

See Also

fitqtl, makeqtl, scanqtl

Examples

data(fake.f2)

# take out several QTLs and make QTL object
qc <- c(1, 8, 13)
qp <- c(26, 56, 28)
fake.f2 <- subset(fake.f2, chr=qc)

fake.f2 <- calc.genoprob(fake.f2, step=2, err=0.001)
qtl <- makeqtl(fake.f2, qc, qp, what="prob")

# fit model with 3 interacting QTLs interacting
# (performing a drop-one-term analysis)
lod <- fitqtl(fake.f2, pheno.col=1, qtl, formula=y~Q1*Q2*Q3,
              method="hk")
summary(lod)

[Package qtl version 1.74 Index]