summaryMap {qtl}	R Documentation

Print summary of a genetic map

Description

Print summary information about a map object.

Usage

## S3 method for class 'map'
summary(object, ...)
summaryMap(object, ...)

Arguments

object
An object of class map, which is a list of vectors (or, for a sex-specific map, 2-row matrices), each specifying the locations of the markers. The object can also be of class cross, in which case the function pull.map is used to extract the genetic map from the object.

...
Ignored at this point.

Value

An object of class summary.map, which is just a data.frame containing the number of markers, length, the average inter-marker spacing, and the maximum distance between markers, for each chromosome and overall. An attribute sexsp indicates whether the map was sex-specific.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

chrlen, pull.map, summary.cross

Examples

data(map10)
summary(map10)

[Package qtl version 1.74 Index]