summary.ripple {qtl}	R Documentation

Print summary of ripple results

Description

Print marker orders, from the output of the function ripple, for which the log10 likelihood relative to the initial order is above a specified cutoff.

Usage

## S3 method for class 'ripple'
summary(object, lod.cutoff = -1, ...)

Arguments

object
An object of class ripple, the output of the function ripple.

lod.cutoff
Only marker orders with LOD score (relative to the initial order) above this cutoff will be displayed. For output of ripple in the case of minimization of the number of obligate crossovers, we double this argument and treat it as a cutoff for the number of obligate crossovers.

...
Ignored at this point.

Value

An object of class summary.ripple, whose rows correspond to marker orders with likelihood (or number of obligate crossovers) within some cutoff of the initial order. If no marker order, other than the initial one, has likelihood within the specified range, the initial and next-best orders are returned.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

ripple, est.map, est.rf

Examples

 
## Not run: data(badorder)
rip1 <- ripple(badorder, 1, 7)
summary(rip1)

rip2 <- ripple(badorder, 1, 2, method="likelihood")
summary(rip2)

badorder <- switch.order(badorder, 1, rip2[2,])

## End(Not run)

[Package qtl version 1.74 Index]