summary.scanPhyloQTL {qtl}	R Documentation

Summarize the results a genome scan to map a QTL to a phylogenetic tree

Description

Print the maximum LOD scores for each partition on each chromosome, from the results of scanPhyloQTL.

Usage

## S3 method for class 'scanPhyloQTL'
summary(object, format=c("postprob", "lod"),
        threshold, ...)

Arguments

object
An object output by the function scanPhyloQTL.

format
Indicates whether to provide LOD scores or approximate posterior probabilities; see Details below.

threshold
A threshold determining which chromosomes should be output; see Details below.

...
Ignored at this point.

Details

This function is used to report chromosomes deemed interesting from a one-QTL genome scan to map QTL to a phylogenetic tree (by scanPhyloQTL).

For format="lod", the output contains the maximum LOD score for each partition on each chromosome (which do not necessarily occur at the same position). The position corresponds to the peak location for the partition with the largest LOD score on that chromosome. The last column is the overall maximum LOD (across partitions) on that chromosome. The second-to-last column is the inferred partition (i.e., that with the largest LOD score. The third-to-last column is the difference between the LOD score for the best partition and that for the second-best.

For format="postprob", the final column contains the maximum LOD score across partitions. But instead of providing the LOD scores for each partition, these are converted to approximate posterior probabilities under the assumption of a single diallelic QTL on that chromosome: on each chromosome, we take 10^LOD for the partitions and rescale them to sum to 1.

The threshold argument is applied to the last column (the maximum LOD score across partitions).

Value

An object of class summary.scanPhyloQTL, to be printed by print.summary.scanPhyloQTL.

Author(s)

Karl W Broman, broman@wisc.edu

References

Broman, K. W., Kim, S., An\'e, C. and Payseur, B. A. Mapping quantitative trait loci to a phylogenetic tree. In preparation.

See Also

scanPhyloQTL, plot.scanPhyloQTL, max.scanPhyloQTL, summary.scanone, inferredpartitions, simPhyloQTL

Examples

## Not run: 
# example map; drop X chromosome
data(map10)
map10 <- map10[1:19]

# simulate data
x <- simPhyloQTL(4, partition="AB|CD", crosses=c("AB", "AC", "AD"),
                 map=map10, n.ind=150,
                 model=c(1, 50, 0.5, 0))

# run calc.genoprob on each cross
x <- lapply(x, calc.genoprob, step=2)

# scan genome, at each position trying all possible partitions
out <- scanPhyloQTL(x, method="hk")

# maximum peak
max(out, format="lod")

# approximate posterior probabilities at peak
max(out, format="postprob")

# all peaks above a threshold for LOD(best) - LOD(2nd best)
summary(out, threshold=1, format="lod")

# all peaks above a threshold for LOD(best), showing approx post'r prob
summary(out, format="postprob", threshold=3)

# plot of results
plot(out)

## End(Not run)

[Package qtl version 1.74 Index]