summary.scanoneboot {qtl}	R Documentation

Bootstrap confidence interval for QTL location

Description

Calculates a bootstrap confidence interval for QTL location, using the bootstrap results from scanoneboot.

Usage

## S3 method for class 'scanoneboot'
summary(object, prob=0.95, expandtomarkers=FALSE, ...)

Arguments

object
Output from scanoneboot.

prob
Desired coverage.

expandtomarkers
If TRUE, the interval is expanded to the nearest flanking markers.

...
Ignored at this point.

Value

An object of class scanone, indicating the position with the maximum LOD, and indicating endpoints for the estimated bootstrap confidence interval.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

scanoneboot, plot.scanoneboot, lodint, bayesint

Examples

## Not run: data(fake.f2)
fake.f2 <- calc.genoprob(fake.f2, step=1, err=0.001)
bootoutput <- scanoneboot(fake.f2, chr=13, method="hk")

summary(bootoutput)
## End(Not run)

[Package qtl version 1.74 Index]