summaryScantwoOld {qtl}	R Documentation

Summarize the results of a two-dimensional genome scan

Description

Summarize the interesting aspects of the results of scantwo; this is the version of summary.scantwo that was included in R/qtl version 1.03 and earlier.

Usage

summaryScantwoOld(object, thresholds = c(0, 0, 0), lodcolumn=1,
                type = c("joint","interaction"), ...)

Arguments

object
An object of class scantwo, the output of the function scantwo.

thresholds
A vector of length three, giving LOD thresholds for the joint LOD, interaction LOD and single-QTL conditional LOD. Negative threshold values are taken relative to the maximum joint, interaction, or individual QTL LOD, respectively.

lodcolumn
If the scantwo results contain LOD scores for multiple phenotypes, this argument indicates which to use in the summary.

type
Indicates whether to pick peaks with maximal joint or interaction LOD.

...
Ignored at this point.

Details

For each pair of chromosomes, the pair of loci for which the LOD score (either joint or interaction LOD, according to the argument type) is a maximum is considered. The pair is printed only if its joint LOD score exceeds the joint threshold and either (a) the interaction LOD score exceeds its threshold or (b) both of the loci have conditional LOD scores that are above the conditional LOD threshold, where the conditional LOD score for locus q1, LOD(q1 | q2), is the log10 likelihood ratio comparing the model with q1 and q2 acting additively to the model with q2 alone.

In the case the results of scanone are not available, the maximum locus pair for each chromosome is printed whenever its joint LOD exceeds the joint LOD threshold.

The criterion used in this summary is due to Gary Churchill and Śaunak Sen, and deserves careful consideration and possible revision.

Value

An object of class summary.scantwo.old, to be printed by print.summary.scantwo.old. Pairs of loci meeting the specified criteria are printed, with their joint LOD, interaction LOD, and the conditional LOD for each locus, along with single-point P-values calculated by the chi-square approximation. P-values are printed as -log10(P).

If the input scantwo object does not include the results of scanone, the interaction and conditional LOD thresholds are ignored, and all pairs of loci for which the joint LOD exceeds its threshold are printed, though without their conditional LOD scores.

Author(s)

Hao Wu; Karl W Broman, broman@wisc.edu; Brian Yandell

See Also

summary.scantwo, scantwo, plot.scantwo, max.scantwo

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2, step=5)
out.2dim <- scantwo(fake.f2, method="hk")

# All pairs of loci
summaryScantwoOld(out.2dim)

# Pairs meeting specified criteria
summaryScantwoOld(out.2dim, c(7, 3, 3))

# Pairs with both conditional LODs > 2
summaryScantwoOld(out.2dim,c(0,1000,2))

# Pairs with interaction LOD is above 3
summaryScantwoOld(out.2dim,c(0,3,1000))

[Package qtl version 1.74 Index]