summary.scantwoperm {qtl}	R Documentation

LOD thresholds from scantwo permutation results

Description

Print the estimated genome-wide LOD thresholds on the basis of permutation results from scantwo (with n.perm > 0).

Usage

## S3 method for class 'scantwoperm'
summary(object, alpha=c(0.05, 0.10), ...)

Arguments

object
Output from the function scantwo with n.perm > 0.

alpha
Genome-wide significance levels.

...
Ignored at this point.

Details

We take the 1-alpha quantiles of the individual LOD scores.

In the case of X-chr-specific permutations, we use the combined length of the autosomes, L_A, and the length of the X chromosome, L_X, and calculate the area of the A:A, A:X, and X:X regions as L_A^2/2, L_A L_X, and L_X^2/2, and then use the nominal significance levels of 1 - (1-alpha)^p, where p is the proportional area for that region.

Value

An object of class summary.scantwoperm, to be printed by print.summary.scantwoperm.

Author(s)

Karl W Broman, broman@wisc.edu

References

Churchill, G. A. and Doerge, R. W. (1994) Empirical threshold values for quantitative trait mapping. Genetics 138, 963–971.

See Also

scantwo, summary.scantwo, plot.scantwoperm

Examples

data(fake.f2)

fake.f2 <- calc.genoprob(fake.f2, step=0)

## Not run: operm <- scantwo(fake.f2, n.perm=100, method="hk")
summary(operm)

[Package qtl version 1.74 Index]