switch.order {qtl}	R Documentation

Switch the order of markers on a chromosome

Description

Switch the order of markers on a specified chromosome to a specified new order.

Usage

switch.order(cross, chr, order, error.prob=0.0001,
                    map.function=c("haldane","kosambi","c-f","morgan"),
                    maxit=4000, tol=1e-6, sex.sp=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
The chromosome for which the marker order is to be switched. Only one chromosome is allowed. (This should be a character string referring to the chromosomes by name.)

order
A vector of numeric indices defining the new marker order. The vector may have length two more than the number of markers, for easy in use with the output of the function ripple.

error.prob
Assumed genotyping error rate (passed to est.map).

map.function
Map function to be used (passed to est.map).

maxit
Maximum number of EM iterations to perform.

tol
Tolerance for determining convergence.

sex.sp
Indicates whether to estimate sex-specific maps; this is used only for the 4-way cross.

Value

The input cross object, but with the marker order on the specified chromosome updated, and with any derived data removed (except for recombination fractions, if present, which are not removed); the genetic map for the relevant chromosome is re-estimated.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

flip.order, ripple, clean.cross

Examples

data(fake.f2)
fake.f2 <- switch.order(fake.f2, 1, c(1,3,2,4:7))

[Package qtl version 1.74 Index]