switchAlleles {qtl}	R Documentation

Switch alleles at selected markers

Description

Switch alleles at selected markers in a cross object.

Usage

switchAlleles(cross, markers, switch=c("AB", "CD", "ABCD", "parents"))

Arguments

cross
An object of class cross. See read.cross for details.

markers
Names of markers whose alleles are to be switched.

switch
For a 4-way cross, indicates how to switch the alleles (A for B, C for D, both A for B and C for D), or both A for C and B for D (parents).

Details

For a backcross, we exchange homozygotes (AA) and heterozygotes (AB).

For doubled haploids and recombinant inbred lines, we exchange the two homozygotes.

For an intercross, we exchange the two homozygotes, and exchange C (i.e., not AA) and D (i.e., not BB). (The heterozygotes in an intercross are left unchanged.)

For a 4-way cross, we consider the argument switch, and the exchanges among the genotypes are more complicated.

Value

The input cross object, with alleles at selected markers switched.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

checkAlleles, est.rf, geno.crosstab

Examples

data(fake.f2)
geno.crosstab(fake.f2, "D5M391", "D5M81")

# switch homozygotes at marker D5M391
fake.f2 <- switchAlleles(fake.f2, "D5M391")

geno.crosstab(fake.f2, "D5M391", "D5M81")

## Not run: fake.f2 <- est.rf(fake.f2)
checkAlleles(fake.f2)

## End(Not run)

[Package qtl version 1.74 Index]