top.errorlod {qtl}	R Documentation

List genotypes with large error LOD scores

Description

Prints those genotypes with error LOD scores above a specified cutoff.

Usage

top.errorlod(cross, chr, cutoff=4, msg=TRUE)

Arguments

cross
An object of class cross. See read.cross for details.

chr
Optional vector indicating the chromosomes to consider. This should be a vector of character strings referring to chromosomes by name; numeric values are converted to strings. Refer to chromosomes with a preceding - to have all chromosomes but those considered. A logical (TRUE/FALSE) vector may also be used.

cutoff
Only those genotypes with error LOD scores above this cutoff will be listed.

msg
If TRUE, print a message if there are no apparent errors.

Value

A data.frame with 4 columns, whose rows correspond to the genotypes that are possibly in error. The four columns give the chromosome number, individual number, marker name, and error LOD score.

Author(s)

Karl W Broman, broman@wisc.edu

See Also

calc.errorlod, plotGeno, plotErrorlod

Examples

data(hyper)

# Calculate error LOD scores
hyper <- calc.errorlod(hyper,error.prob=0.01)

# Print those above a specified cutoff
top.errorlod(hyper,cutoff=4)

[Package qtl version 1.74 Index]