transformPheno {qtl}	R Documentation

Transformation of the phenotypes in a cross object

Description

Transform phenotypes in a cross object; by default use a logarithmic transformation, though any function may be used.

Usage

transformPheno(cross, pheno.col=1, transf=log, ...)

Arguments

cross
An object of class cross. See read.cross for details.

pheno.col
A vector of numeric indices or character strings (indicating phenotypes by name) of phenotypes to be transformed.

transf
The function to use in the transformation.

...
Additional arguments, to be passed to transf.

Value

The input cross object with the transformed phenotypes

Author(s)

Danny Arends danny.arends@gmail.com

See Also

mqmscan, scanone

Examples

data(multitrait)

# Log transformation of all phenotypes
multitrait.log <- transformPheno(multitrait, pheno.col=1:nphe(multitrait))

# Square-root transformation of all phenotypes
multitrait.sqrt <- transformPheno(multitrait, pheno.col=1:nphe(multitrait),
                                  transf=sqrt)

[Package qtl version 1.74 Index]